+++ internal/audio/audiofile/audiofile.go
@@ -1,7 +1,6 @@
 package audiofile
 
 import (
-	"bufio"
 	"errors"
 	"fmt"
 	"iter"
@@ -11,7 +10,6 @@ import (
 	"regexp"
 	"slices"
 	"strconv"
-	"strings"
 	"talkliketv.click/tltv/internal/models"
 	"talkliketv.click/tltv/internal/util"
 
@@ -82,136 +80,28 @@ func (r *RealCmdRunner) CombinedOutput(cmd *exec.Cmd) ([]byte, error) {
 // line and then parses the file accordingly, returning a string slice containing the
 // phrases to be translated
 func (af *AudioFile) GetLines(e echo.Context, f multipart.File) ([]string, error) {
-	// get file type, options are srt, single line text or paragraph
-	fileType := ""
-	scanner := bufio.NewScanner(f)
-	//start at the first line again
-	_, err := f.Seek(0, 0)
+	fileType, err := detectFileType(f)
 	if err != nil {
 		e.Logger().Error(err)
 		return nil, err
 	}
-	count := 0
-	var line string
 
-	// verify if file is srt
-	for scanner.Scan() {
-		if fileType != "" || count > 5 {
-			break
-		}
-		line = scanner.Text()
-		// if line contains ">" and doesn't contain any letters it is srt file
-		if strings.Contains(line, ">") {
-			if strings.Contains(line, "<font") || !reAlpha.MatchString(line) {
-				fileType = "srt"
-			}
-		}
-		count++
-	}
-	//start at the first line again
-	_, err = f.Seek(0, 0)
-	if err != nil {
-		e.Logger().Error(err)
+	// Reset file pointer again
+	if _, err := f.Seek(0, 0); err != nil {
 		return nil, err
 	}
-	count = 0
-	scanner = bufio.NewScanner(f)
-	// verify if file is in paragraph form
-	for scanner.Scan() {
-		if fileType != "" || count > 4 {
-			break
-		}
-		line = scanner.Text()
-		// Split on punctuation characters
-		re := regexp.MustCompile(`[.!?]`)
-		result := re.Split(line, -1)
-		if len(result) > 3 {
-			fileType = "paragraph"
-		}
-		count++
-	}
-	// TODO verify single phrase per line form (these can be multiple sentences per line)
-	_, err = f.Seek(0, 0)
-	if err != nil {
-		e.Logger().Error(err)
-		return nil, err
-	}
-	var stringsSlice []string
-	if fileType == "srt" {
-		stringsSlice = parseSrt(f)
-	}
-	if fileType == "paragraph" {
-		stringsSlice = parseParagraph(f)
-	}
-	if fileType == "" {
-		stringsSlice = parseSingle(f)
-	}
-	if len(stringsSlice) == 0 {
-		return nil, errors.New("unable to parsefile file")
+
+	lines, err := parseFileContent(f, fileType)
+	if err != nil || len(lines) == 0 {
+		return nil, errors.New("unable to parse file")
 	}
 
-	// remove strings longer than 150 characters and duplicates from stringsSlice
-	return util.RemoveLongStr(util.RemoveDuplicateStr(stringsSlice)), nil
-}
-
-// CreateMp3Zip takes the input txt files created with BuildAudioInputFiles and uses ffmpeg
-// to build an output mp3's file and the zips them into a single file to be returned to the
-// requester
-func (af *AudioFile) CreateMp3Zip(e echo.Context, t models.Title, tmpDir string) (*os.File, error) {
-	// get a list of files from the temp directory
-	files, err := os.ReadDir(tmpDir)
-	if err != nil {
-		e.Logger().Error(err)
-		return nil, err
-	}
-	if len(files) == 0 {
-		return nil, errors.New("no files found in CreateMp3Zip")
-	}
-	// create outputs folder to hold all the mp3's to zip
-	outDirPath := tmpDir + "outputs"
-	err = os.MkdirAll(outDirPath, 0777)
-	if err != nil {
-		e.Logger().Error(err)
-		return nil, err
-	}
-	for i, f := range files {
-		// ffmpeg -f concat -safe 0 -i ffmpeg_input.txt -c copy output.mp3
-		outputString := fmt.Sprintf("%s/%s-%d.mp3", outDirPath, t.Name, i)
-		cmd := exec.Command("ffmpeg", "-f", "concat", "-safe", "0", "-i", tmpDir+f.Name(), "-c", "copy", outputString) //nolint:gosec
-
-		//Execute the command and get the output
-		output, err := af.cmdX.CombinedOutput(cmd)
-		if err != nil {
-			e.Logger().Error(err)
-			e.Logger().Error("combined output: " + string(output))
-			return nil, err
-		}
-	}
-
-	// add a text files of the translated phrases, this is useful studying
-	if t.ToPhrases != nil {
-		// Create file to write all the translated phrases to
-		f, err := os.Create(outDirPath + "/" + t.Name + "-translates.txt")
-		if err != nil {
-			e.Logger().Error(err)
-			return nil, err
-		}
-		defer f.Close()
-		for _, text := range t.ToPhrases {
-			_, err = f.WriteString(text.Text + "\n")
-			if err != nil {
-				e.Logger().Error(err)
-				return nil, err
-			}
-		}
-	}
-
-	return util.CreateZipFile(e, tmpDir, t.Name, outDirPath)
+	return util.RemoveLongStr(util.RemoveDuplicateStr(lines)), nil
 }
 
 // BuildAudioInputFiles creates a file with the filepaths of the mp3's used to construct
 // the output files with ffmpeg in CreateMp3Zip
-func (af *AudioFile) BuildAudioInputFiles(e echo.Context, t models.Title, pause, from, to, tmpDir string) error {
+func (af *AudioFile) BuildAudioInputFiles(e echo.Context, t models.Title, pause, fromLang, toLang, tmpDir string) error {
 	maxP := len(t.TitlePhrases) - 1
 
 	pattern := audio.GetPattern(t.Pattern)
@@ -232,7 +122,6 @@ func (af *AudioFile) BuildAudioInputFiles(e echo.Context, t models.Title, pause,
 			e.Logger().Error(err)
 			return err
 		}
-		defer f.Close()
 
 		// start audiofile with silence
 		_, err = f.WriteString(fmt.Sprintf("file '%s'\n", pause))
@@ -240,58 +129,42 @@ func (af *AudioFile) BuildAudioInputFiles(e echo.Context, t models.Title, pause,
 			e.Logger().Error(err)
 			return err
 		}
-		for _, audioFloat := range chunk {
-			// convert float representation of phrase id and whether speech should be native or translated
-			// it is represented in /internal/pattern as "phraseId"."native_boolean" as a float32 to save space
-			stringFloat := strconv.FormatFloat(float64(audioFloat), 'f', -1, 32)
-			phraseNative := strings.Split(stringFloat, ".")
-			// if when you split the string the length is 1 that means the float ended in .0 which means audio
-			// should be translated; if length is 2 that means float ended in .1 which indicates it should be
-			// native
-			native := false
-			if len(phraseNative) == 2 {
-				native = true
-			}
-			phraseId, err := strconv.Atoi(phraseNative[0])
+		for _, audioTok := range chunk {
+			phraseIdKey, nativeLang, err := SplitShortString(strconv.Itoa(int(audioTok)))
 			if err != nil {
 				e.Logger().Error(err)
 				return err
 			}
+			native := false
+			if nativeLang == "1" {
+				native = true
+			}
 			// if: we have reached the highest phrase id then this will be the last audio block
 			// else if: skip if phraseId does not exist (is greater than maxP)
-			// else if: native language then we add filepath for from audio mp3
-			// else: add audio filepath for language you want to learn
+			// else if: native language then we add filepath for from language audio mp3
+			// else: add audio filepath for to language mp3
+			phraseId, err := strconv.Atoi(phraseIdKey)
+			if err != nil {
+				e.Logger().Error(err)
+				return err
+			}
 			if phraseId == maxP {
 				last = true
 			}
-			if native {
-				_, err = f.WriteString(fmt.Sprintf("file '%s%d'\n", from, phraseId))
-				if err != nil {
-					e.Logger().Error(err)
-					return err
-				}
-				_, err = f.WriteString(fmt.Sprintf("file '%s'\n", pause))
-				if err != nil {
-					e.Logger().Error(err)
-					return err
-				}
-			} else {
-				_, err = f.WriteString(fmt.Sprintf("file '%s%d'\n", to, phraseId))
-				if err != nil {
-					e.Logger().Error(err)
-					return err
-				}
-				_, err = f.WriteString(fmt.Sprintf("file '%s'\n", pause))
-				if err != nil {
-					e.Logger().Error(err)
-					return err
-				}
+
+			if err = writeStringToFile(e, native, f, fromLang, toLang, phraseIdKey, pause); err != nil {
+				return err
 			}
 		}
 		// end audiofile with silence
 		_, err = f.WriteString(fmt.Sprintf("file '%s'\n", pause))
 		if err != nil {
 			e.Logger().Error(err)
+			return err
+		}
+		// Close the file explicitly
+		if err = f.Close(); err != nil {
+			e.Logger().Error("failed to close file: %v", err)
 			return err
 		}
 		if last {
@@ -302,35 +175,34 @@ func (af *AudioFile) BuildAudioInputFiles(e echo.Context, t models.Title, pause,
 	return nil
 }
 
-// CreatePhrasesZip creates a zipped file of txt files from the file the user uploaded if it contains
-// more phrases than the limit of config.MaxNumPhrases. It takes a iter.Seq of strings and outputs them
-// to files, each chunk containing config.MaxNumPhrases and than zips them up. Sending them back to the
-// user
-func (af *AudioFile) CreatePhrasesZip(e echo.Context, chunkedPhrases iter.Seq[[]string], tmpPath, filename string) (*os.File, error) {
-	// create outputs folder to hold all the txt files to zip
-	err := os.MkdirAll(tmpPath, 0777)
+func writeStringToFile(e echo.Context, native bool, f *os.File, fromLang, toLang, phraseIdKey, pause string) error {
+	audioString := ""
+	if native {
+		audioString = fmt.Sprintf("file '%s%s'\n", fromLang, phraseIdKey)
+	} else {
+		audioString = fmt.Sprintf("file '%s%s'\n", toLang, phraseIdKey)
+	}
+
+	pauseString := fmt.Sprintf("file '%s'\n", pause)
+	_, err := f.WriteString(audioString + pauseString)
 	if err != nil {
 		e.Logger().Error(err)
-		return nil, err
-	}
-	count := 0
-	for chunk := range chunkedPhrases {
-		file := fmt.Sprintf("%s-phrases-%d.txt", filename, count)
-		count++
-		f, err := os.Create(tmpPath + file)
-		if err != nil {
-			e.Logger().Error(err)
-			return nil, err
-		}
-		defer f.Close()
-
-		for _, phrase := range chunk {
-			_, err = f.WriteString(phrase + "\n")
-			if err != nil {
-				return nil, err
-			}
-		}
+		return err
 	}
 
-	return util.CreateZipFile(e, tmpPath, filename, tmpPath)
+	return nil
+}
+
+// SplitShortString splits a string into two parts: the first part contains all but the last character,
+// the last character represents a bool indicating whether the audio input should be the native language,
+// the first part is the phraseId
+func SplitShortString(input string) (string, string, error) {
+	if len(input) < 2 {
+		return "", "", errors.New("input string must have at least two characters")
+	}
+
+	lastDigit := (input)[len(input)-1:]
+	remainingDigits := (input)[:len(input)-1]
+
+	return remainingDigits, lastDigit, nil
 }
