+++ internal/audio/pattern/pattern.go
@@ -1,6 +1,6 @@
 package audio
 
-func GetPattern(i int) []float32 {
+func GetPattern(i int) []uint16 {
 	switch i {
 	case 1:
 		return Intermediate
@@ -19,11 +19,10 @@ func GetPattern(i int) []float32 {
 // Frequency of repetition can be adjusted in the program
 // This pattern is set with multi = 15 at https://github.com/dsaker/audio-pattern-builder/blob/c577c22826aad34906f4eb944ff92c9693f9b4c4/main.go#L26
 // Intermediate multi = 15
-var Intermediate = []float32{0.1, 0.0, 0.0, 1.1, 1.0, 1.0, 2.1, 2.0, 2.0, 3.1, 3.0, 3.0, 4.1, 4.0, 4.0, 5.1, 5.0, 5.0, 0.1, 0.0, 6.1, 6.0, 6.0, 1.1, 1.0, 7.1, 7.0, 7.0, 2.1, 2.0, 8.1, 8.0, 8.0, 3.1, 3.0, 9.1, 9.0, 9.0, 4.1, 4.0, 5.1, 5.0, 10.1, 10.0, 10.0, 6.1, 6.0, 11.1, 11.0, 11.0, 7.1, 7.0, 12.1, 12.0, 12.0, 8.1, 8.0, 13.1, 13.0, 13.0, 9.1, 9.0, 14.1, 14.0, 14.0, 10.1, 10.0, 15.1, 15.0, 15.0, 11.1, 11.0, 16.1, 16.0, 16.0, 12.1, 12.0, 17.1, 17.0, 17.0, 13.1, 13.0, 18.1, 18.0, 18.0, 14.1, 14.0, 19.1, 19.0, 19.0, 15.1, 15.0, 20.1, 20.0, 20.0, 16.1, 16.0, 21.1, 21.0, 21.0, 17.1, 17.0, 22.1, 22.0, 22.0, 18.1, 18.0, 0.1, 0.0, 19.1, 19.0, 1.1, 1.0, 20.1, 20.0, 2.1, 2.0, 21.1, 21.0, 3.1, 3.0, 22.1, 22.0, 4.1, 4.0, 5.1, 5.0, 23.1, 23.0, 23.0, 6.1, 6.0, 24.1, 24.0, 24.0, 7.1, 7.0, 25.1, 25.0, 25.0, 8.1, 8.0, 26.1, 26.0, 26.0, 9.1, 9.0, 27.1, 27.0, 27.0, 10.1, 10.0, 23.1, 23.0, 11.1, 11.0, 24.1, 24.0, 12.1, 12.0, 25.1, 25.0, 13.1, 13.0, 26.1, 26.0, 27.1, 27.0, 14.1, 14.0, 28.1, 28.0, 28.0, 15.1, 15.0, 29.1, 29.0, 29.0, 16.1, 16.0, 30.1, 30.0, 30.0, 17.1, 17.0, 31.1, 31.0, 31.0, 18.1, 18.0, 32.1, 32.0, 32.0, 19.1, 19.0, 28.1, 28.0, 20.1, 20.0, 29.1, 29.0, 21.1, 21.0, 30.1, 30.0, 22.1, 22.0, 31.1, 31.0, 32.1, 32.0, 33.1, 33.0, 33.0, 34.1, 34.0, 34.0, 35.1, 35.0, 35.0, 36.1, 36.0, 36.0, 37.1, 37.0, 37.0, 38.1, 38.0, 38.0, 23.1, 23.0, 33.1, 33.0, 24.1, 24.0, 34.1, 34.0, 25.1, 25.0, 35.1, 35.0, 26.1, 26.0, 27.1, 27.0, 36.1, 36.0, 37.1, 37.0, 38.1, 38.0, 39.1, 39.0, 39.0, 40.1, 40.0, 40.0, 41.1, 41.0, 41.0, 42.1, 42.0, 42.0, 28.1, 28.0, 43.1, 43.0, 43.0, 29.1, 29.0, 44.1, 44.0, 44.0, 30.1, 30.0, 39.1, 39.0, 31.1, 31.0, 32.1, 32.0, 0.1, 0.0, 40.1, 40.0, 1.1, 1.0, 41.1, 41.0, 2.1, 2.0, 42.1, 42.0, 3.1, 3.0, 43.1, 43.0, 4.1, 4.0, 5.1, 5.0, 44.1, 44.0, 6.1, 6.0, 33.1, 33.0, 7.1, 7.0, 34.1, 34.0, 8.1, 8.0, 35.1, 35.0, 9.1, 9.0, 36.1, 36.0, 37.1, 37.0, 10.1, 10.0, 38.1, 38.0, 11.1, 11.0, 45.1, 45.0, 45.0, 12.1, 12.0, 46.1, 46.0, 46.0, 13.1, 13.0, 47.1, 47.0, 47.0, 14.1, 14.0, 48.1, 48.0, 48.0, 15.1, 15.0, 39.1, 39.0, 16.1, 16.0, 45.1, 45.0, 17.1, 17.0, 40.1, 40.0, 18.1, 18.0, 41.1, 41.0, 19.1, 19.0, 42.1, 42.0, 20.1, 20.0, 43.1, 43.0, 21.1, 21.0, 44.1, 44.0, 22.1, 22.0, 46.1, 46.0, 47.1, 47.0, 48.1, 48.0, 49.1, 49.0, 49.0, 50.1, 50.0, 50.0, 51.1, 51.0, 51.0, 52.1, 52.0, 52.0, 53.1, 53.0, 53.0, 23.1, 23.0, 54.1, 54.0, 54.0, 24.1, 24.0, 49.1, 49.0, 25.1, 25.0, 50.1, 50.0, 26.1, 26.0, 27.1, 27.0, 51.1, 51.0, 45.1, 45.0, 52.1, 52.0, 53.1, 53.0, 54.1, 54.0, 55.1, 55.0, 55.0, 56.1, 56.0, 56.0, 28.1, 28.0, 57.1, 57.0, 57.0, 29.1, 29.0, 58.1, 58.0, 58.0, 30.1, 30.0, 59.1, 59.0, 59.0, 31.1, 31.0, 32.1, 32.0, 46.1, 46.0, 47.1, 47.0, 48.1, 48.0, 55.1, 55.0, 56.1, 56.0, 57.1, 57.0, 58.1, 58.0, 59.1, 59.0, 60.1, 60.0, 60.0, 33.1, 33.0, 61.1, 61.0, 61.0, 34.1, 34.0, 49.1, 49.0, 35.1, 35.0, 50.1, 50.0, 36.1, 36.0, 37.1, 37.0, 38.1, 38.0, 51.1, 51.0, 52.1, 52.0, 53.1, 53.0, 54.1, 54.0, 60.1, 60.0, 61.1, 61.0, 62.1, 62.0, 62.0, 39.1, 39.0, 63.1, 63.0, 63.0, 64.1, 64.0, 64.0, 40.1, 40.0, 65.1, 65.0, 65.0, 41.1, 41.0, 66.1, 66.0, 66.0, 42.1, 42.0, 62.1, 62.0, 43.1, 43.0, 55.1, 55.0, 44.1, 44.0, 56.1, 56.0, 57.1, 57.0, 58.1, 58.0, 59.1, 59.0, 63.1, 63.0, 64.1, 64.0, 65.1, 65.0, 66.1, 66.0, 67.1, 67.0, 67.0, 68.1, 68.0, 68.0, 69.1, 69.0, 69.0, 70.1, 70.0, 70.0, 71.1, 71.0, 71.0, 60.1, 60.0, 61.1, 61.0, 0.1, 0.0, 67.1, 67.0, 1.1, 1.0, 45.1, 45.0, 2.1, 2.0, 68.1, 68.0, 3.1, 3.0, 69.1, 69.0, 4.1, 4.0, 5.1, 5.0, 62.1, 62.0, 6.1, 6.0, 70.1, 70.0, 7.1, 7.0, 71.1, 71.0, 8.1, 8.0, 72.1, 72.0, 72.0, 9.1, 9.0, 73.1, 73.0, 73.0, 46.1, 46.0, 10.1, 10.0, 47.1, 47.0, 11.1, 11.0, 48.1, 48.0, 12.1, 12.0, 63.1, 63.0, 13.1, 13.0, 64.1, 64.0, 65.1, 65.0, 14.1, 14.0, 66.1, 66.0, 15.1, 15.0, 49.1, 49.0, 16.1, 16.0, 50.1, 50.0, 17.1, 17.0, 67.1, 67.0, 18.1, 18.0, 51.1, 51.0, 19.1, 19.0, 52.1, 52.0, 20.1, 20.0, 53.1, 53.0, 21.1, 21.0, 54.1, 54.0, 22.1, 22.0, 68.1, 68.0, 69.1, 69.0, 70.1, 70.0, 71.1, 71.0, 72.1, 72.0, 73.1, 73.0, 74.1, 74.0, 74.0, 55.1, 55.0, 75.1, 75.0, 75.0, 56.1, 56.0, 23.1, 23.0, 57.1, 57.0, 24.1, 24.0, 58.1, 58.0, 25.1, 25.0, 59.1, 59.0, 26.1, 26.0, 27.1, 27.0, 74.1, 74.0, 75.1, 75.0, 76.1, 76.0, 76.0, 77.1, 77.0, 77.0, 78.1, 78.0, 78.0, 60.1, 60.0, 61.1, 61.0, 28.1, 28.0, 79.1, 79.0, 79.0, 29.1, 29.0, 76.1, 76.0, 30.1, 30.0, 77.1, 77.0, 31.1, 31.0, 32.1, 32.0, 78.1, 78.0, 72.1, 72.0, 62.1, 62.0, 73.1, 73.0, 79.1, 79.0, 80.1, 80.0, 80.0, 81.1, 81.0, 81.0, 82.1, 82.0, 82.0, 33.1, 33.0, 83.1, 83.0, 83.0, 34.1, 34.0, 84.1, 84.0, 84.0, 35.1, 35.0, 80.1, 80.0, 36.1, 36.0, 37.1, 37.0, 38.1, 38.0, 63.1, 63.0, 64.1, 64.0, 65.1, 65.0, 66.1, 66.0, 74.1, 74.0, 75.1, 75.0, 81.1, 81.0, 82.1, 82.0, 39.1, 39.0, 67.1, 67.0, 76.1, 76.0, 77.1, 77.0, 40.1, 40.0, 78.1, 78.0, 41.1, 41.0, 83.1, 83.0, 42.1, 42.0, 84.1, 84.0, 43.1, 43.0, 79.1, 79.0, 44.1, 44.0, 68.1, 68.0, 69.1, 69.0, 70.1, 70.0, 71.1, 71.0, 85.1, 85.0, 85.0, 86.1, 86.0, 86.0, 80.1, 80.0, 87.1, 87.0, 87.0, 88.1, 88.0, 88.0, 89.1, 89.0, 89.0, 90.1, 90.0, 90.0, 85.1, 85.0, 86.1, 86.0, 91.1, 91.0, 91.0, 87.1, 87.0, 81.1, 81.0, 82.1, 82.0, 88.1, 88.0, 45.1, 45.0, 89.1, 89.0, 90.1, 90.0, 91.1, 91.0, 92.1, 92.0, 92.0, 83.1, 83.0, 93.1, 93.0, 93.0, 84.1, 84.0, 94.1, 94.0, 94.0, 95.1, 95.0, 95.0, 96.1, 96.0, 96.0, 92.1, 92.0, 46.1, 46.0, 72.1, 72.0, 47.1, 47.0, 73.1, 73.0, 48.1, 48.0, 93.1, 93.0, 94.1, 94.0, 95.1, 95.0, 96.1, 96.0, 97.1, 97.0, 97.0, 85.1, 85.0, 49.1, 49.0, 86.1, 86.0, 50.1, 50.0, 87.1, 87.0, 98.1, 98.0, 98.0, 88.1, 88.0, 51.1, 51.0, 89.1, 89.0, 52.1, 52.0, 90.1, 90.0, 53.1, 53.0, 91.1, 91.0, 54.1, 54.0, 74.1, 74.0, 75.1, 75.0, 97.1, 97.0, 98.1, 98.0, 99.1, 99.0, 99.0, 92.1, 92.0, 76.1, 76.0, 77.1, 77.0, 78.1, 78.0, 55.1, 55.0, 100.1, 100.0, 100.0, 56.1, 56.0, 93.1, 93.0, 57.1, 57.0, 79.1, 79.0, 58.1, 58.0, 94.1, 94.0, 59.1, 59.0, 95.1, 95.0, 96.1, 96.0, 99.1, 99.0, 100.1, 100.0, 80.1, 80.0, 101.1, 101.0, 101.0, 102.1, 102.0, 102.0, 103.1, 103.0, 103.0, 60.1, 60.0, 61.1, 61.0, 104.1, 104.0, 104.0, 105.1, 105.0, 105.0, 101.1, 101.0, 81.1, 81.0, 82.1, 82.0, 97.1, 97.0, 0.1, 0.0, 98.1, 98.0, 1.1, 1.0, 62.1, 62.0, 2.1, 2.0, 102.1, 102.0, 3.1, 3.0, 83.1, 83.0, 4.1, 4.0, 5.1, 5.0, 84.1, 84.0, 6.1, 6.0, 103.1, 103.0, 7.1, 7.0, 104.1, 104.0, 8.1, 8.0, 105.1, 105.0, 9.1, 9.0, 99.1, 99.0, 100.1, 100.0, 10.1, 10.0, 63.1, 63.0, 11.1, 11.0, 64.1, 64.0, 12.1, 12.0, 65.1, 65.0, 13.1, 13.0, 66.1, 66.0, 85.1, 85.0, 14.1, 14.0, 67.1, 67.0, 15.1, 15.0, 86.1, 86.0, 16.1, 16.0, 87.1, 87.0, 17.1, 17.0, 88.1, 88.0, 18.1, 18.0, 89.1, 89.0, 19.1, 19.0, 90.1, 90.0, 20.1, 20.0, 68.1, 68.0, 21.1, 21.0, 69.1, 69.0, 22.1, 22.0, 70.1, 70.0, 71.1, 71.0, 91.1, 91.0, 92.1, 92.0, 93.1, 93.0, 101.1, 101.0, 94.1, 94.0, 102.1, 102.0, 95.1, 95.0, 96.1, 96.0, 23.1, 23.0, 103.1, 103.0, 24.1, 24.0, 104.1, 104.0, 25.1, 25.0, 105.1, 105.0, 26.1, 26.0, 27.1, 27.0, 106.1, 106.0, 106.0, 107.1, 107.0, 107.0, 108.1, 108.0, 108.0, 109.1, 109.0, 109.0, 110.1, 110.0, 110.0, 97.1, 97.0, 111.1, 111.0, 111.0, 28.1, 28.0, 98.1, 98.0, 29.1, 29.0, 106.1, 106.0, 30.1, 30.0, 72.1, 72.0, 31.1, 31.0, 32.1, 32.0, 73.1, 73.0, 107.1, 107.0, 108.1, 108.0, 109.1, 109.0, 110.1, 110.0, 111.1, 111.0, 112.1, 112.0, 112.0, 99.1, 99.0, 33.1, 33.0, 100.1, 100.0, 34.1, 34.0, 113.1, 113.0, 113.0, 35.1, 35.0, 114.1, 114.0, 114.0, 36.1, 36.0, 37.1, 37.0, 38.1, 38.0, 74.1, 74.0, 75.1, 75.0, 112.1, 112.0, 113.1, 113.0, 114.1, 114.0, 76.1, 76.0, 77.1, 77.0, 78.1, 78.0, 39.1, 39.0, 106.1, 106.0, 115.1, 115.0, 115.0, 40.1, 40.0, 79.1, 79.0, 41.1, 41.0, 107.1, 107.0, 42.1, 42.0, 108.1, 108.0, 43.1, 43.0, 109.1, 109.0, 44.1, 44.0, 80.1, 80.0, 110.1, 110.0, 111.1, 111.0, 115.1, 115.0, 101.1, 101.0, 116.1, 116.0, 116.0, 102.1, 102.0, 117.1, 117.0, 117.0, 118.1, 118.0, 118.0, 103.1, 103.0, 81.1, 81.0, 82.1, 82.0, 104.1, 104.0, 105.1, 105.0, 112.1, 112.0, 113.1, 113.0, 114.1, 114.0, 116.1, 116.0, 117.1, 117.0, 83.1, 83.0, 45.1, 45.0, 118.1, 118.0, 84.1, 84.0, 119.1, 119.0, 119.0, 120.1, 120.0, 120.0, 121.1, 121.0, 121.0, 122.1, 122.0, 122.0, 123.1, 123.0, 123.0, 124.1, 124.0, 124.0, 119.1, 119.0, 125.1, 125.0, 125.0, 46.1, 46.0, 115.1, 115.0, 47.1, 47.0, 120.1, 120.0, 48.1, 48.0, 121.1, 121.0, 85.1, 85.0, 122.1, 122.0, 123.1, 123.0, 124.1, 124.0, 86.1, 86.0, 125.1, 125.0, 87.1, 87.0, 49.1, 49.0, 88.1, 88.0, 50.1, 50.0, 89.1, 89.0, 90.1, 90.0, 116.1, 116.0, 51.1, 51.0, 117.1, 117.0, 52.1, 52.0, 118.1, 118.0, 53.1, 53.0, 126.1, 126.0, 126.0, 54.1, 54.0, 127.1, 127.0, 127.0, 91.1, 91.0, 92.1, 92.0, 93.1, 93.0, 106.1, 106.0, 94.1, 94.0, 119.1, 119.0, 95.1, 95.0, 55.1, 55.0, 96.1, 96.0, 56.1, 56.0, 107.1, 107.0, 57.1, 57.0, 108.1, 108.0, 58.1, 58.0, 109.1, 109.0, 59.1, 59.0, 110.1, 110.0, 111.1, 111.0, 120.1, 120.0, 121.1, 121.0, 122.1, 122.0, 123.1, 123.0, 124.1, 124.0, 97.1, 97.0, 125.1, 125.0, 60.1, 60.0, 61.1, 61.0, 98.1, 98.0, 112.1, 112.0, 113.1, 113.0, 114.1, 114.0, 126.1, 126.0, 127.1, 127.0, 128.1, 128.0, 128.0, 129.1, 129.0, 129.0, 62.1, 62.0, 130.1, 130.0, 130.0, 131.1, 131.0, 131.0, 99.1, 99.0, 132.1, 132.0, 132.0, 100.1, 100.0, 128.1, 128.0, 129.1, 129.0, 133.1, 133.0, 133.0, 130.1, 130.0, 134.1, 134.0, 134.0, 115.1, 115.0, 131.1, 131.0, 132.1, 132.0, 63.1, 63.0, 135.1, 135.0, 135.0, 64.1, 64.0, 133.1, 133.0, 65.1, 65.0, 134.1, 134.0, 66.1, 66.0, 136.1, 136.0, 136.0, 67.1, 67.0, 135.1, 135.0, 137.1, 137.0, 137.0, 138.1, 138.0, 138.0, 116.1, 116.0, 126.1, 126.0, 117.1, 117.0, 127.1, 127.0, 118.1, 118.0, 136.1, 136.0, 68.1, 68.0, 137.1, 137.0, 69.1, 69.0, 138.1, 138.0, 70.1, 70.0, 71.1, 71.0, 101.1, 101.0, 128.1, 128.0, 102.1, 102.0, 119.1, 119.0, 129.1, 129.0, 103.1, 103.0, 130.1, 130.0, 131.1, 131.0, 104.1, 104.0, 105.1, 105.0, 132.1, 132.0, 133.1, 133.0, 134.1, 134.0, 139.1, 139.0, 139.0, 140.1, 140.0, 140.0, 120.1, 120.0, 121.1, 121.0, 122.1, 122.0, 123.1, 123.0, 124.1, 124.0, 135.1, 135.0, 125.1, 125.0, 139.1, 139.0, 140.1, 140.0, 136.1, 136.0, 141.1, 141.0, 141.0, 137.1, 137.0, 142.1, 142.0, 142.0, 138.1, 138.0, 72.1, 72.0, 143.1, 143.0, 143.0, 144.1, 144.0, 144.0, 73.1, 73.0, 141.1, 141.0, 145.1, 145.0, 145.0, 142.1, 142.0, 146.1, 146.0, 146.0, 147.1, 147.0, 147.0, 143.1, 143.0, 144.1, 144.0, 148.1, 148.0, 148.0, 145.1, 145.0, 149.1, 149.0, 149.0, 146.1, 146.0, 147.1, 147.0, 0.1, 0.0, 150.1, 150.0, 150.0, 1.1, 1.0, 74.1, 74.0, 2.1, 2.0, 75.1, 75.0, 3.1, 3.0, 139.1, 139.0, 4.1, 4.0, 5.1, 5.0, 76.1, 76.0, 6.1, 6.0, 77.1, 77.0, 7.1, 7.0, 78.1, 78.0, 8.1, 8.0, 79.1, 79.0, 9.1, 9.0, 106.1, 106.0, 107.1, 107.0, 10.1, 10.0, 108.1, 108.0, 11.1, 11.0, 80.1, 80.0, 12.1, 12.0, 109.1, 109.0, 13.1, 13.0, 110.1, 110.0, 111.1, 111.0, 14.1, 14.0, 126.1, 126.0, 15.1, 15.0, 127.1, 127.0, 16.1, 16.0, 81.1, 81.0, 17.1, 17.0, 82.1, 82.0, 18.1, 18.0, 112.1, 112.0, 19.1, 19.0, 113.1, 113.0, 20.1, 20.0, 114.1, 114.0, 21.1, 21.0, 83.1, 83.0, 22.1, 22.0, 128.1, 128.0, 84.1, 84.0, 129.1, 129.0, 130.1, 130.0, 131.1, 131.0, 132.1, 132.0, 133.1, 133.0, 115.1, 115.0, 134.1, 134.0, 135.1, 135.0, 136.1, 136.0, 23.1, 23.0, 137.1, 137.0, 24.1, 24.0, 138.1, 138.0, 25.1, 25.0, 140.1, 140.0, 26.1, 26.0, 27.1, 27.0, 85.1, 85.0, 141.1, 141.0, 142.1, 142.0, 86.1, 86.0, 87.1, 87.0, 116.1, 116.0, 88.1, 88.0, 117.1, 117.0, 89.1, 89.0, 28.1, 28.0, 90.1, 90.0, 29.1, 29.0, 118.1, 118.0, 30.1, 30.0, 139.1, 139.0, 31.1, 31.0, 32.1, 32.0, 143.1, 143.0, 91.1, 91.0, 92.1, 92.0, 93.1, 93.0, 119.1, 119.0, 94.1, 94.0, 144.1, 144.0, 95.1, 95.0, 145.1, 145.0, 33.1, 33.0, 96.1, 96.0, 34.1, 34.0, 146.1, 146.0, 35.1, 35.0, 120.1, 120.0, 36.1, 36.0, 37.1, 37.0, 38.1, 38.0, 121.1, 121.0, 122.1, 122.0, 123.1, 123.0, 124.1, 124.0, 125.1, 125.0, 97.1, 97.0, 147.1, 147.0, 148.1, 148.0, 39.1, 39.0, 149.1, 149.0, 98.1, 98.0, 150.1, 150.0, 40.1, 40.0, 151.1, 151.0, 151.0, 41.1, 41.0, 152.1, 152.0, 152.0, 42.1, 42.0, 153.1, 153.0, 153.0, 43.1, 43.0, 154.1, 154.0, 154.0, 44.1, 44.0, 155.1, 155.0, 155.0, 151.1, 151.0, 156.1, 156.0, 156.0, 99.1, 99.0, 152.1, 152.0, 100.1, 100.0, 153.1, 153.0, 154.1, 154.0, 157.1, 157.0, 157.0, 155.1, 155.0, 158.1, 158.0, 158.0, 156.1, 156.0, 159.1, 159.0, 159.0, 160.1, 160.0, 160.0, 140.1, 140.0, 161.1, 161.0, 161.0, 157.1, 157.0, 162.1, 162.0, 162.0, 158.1, 158.0, 141.1, 141.0, 142.1, 142.0, 148.1, 148.0, 45.1, 45.0, 149.1, 149.0, 150.1, 150.0, 159.1, 159.0, 160.1, 160.0, 161.1, 161.0, 162.1, 162.0, 163.1, 163.0, 163.0, 164.1, 164.0, 164.0, 165.1, 165.0, 165.0, 151.1, 151.0, 46.1, 46.0, 143.1, 143.0, 47.1, 47.0, 152.1, 152.0, 48.1, 48.0, 101.1, 101.0, 102.1, 102.0, 126.1, 126.0, 127.1, 127.0, 103.1, 103.0, 144.1, 144.0, 49.1, 49.0, 104.1, 104.0, 50.1, 50.0, 105.1, 105.0, 145.1, 145.0, 146.1, 146.0, 51.1, 51.0, 153.1, 153.0, 52.1, 52.0, 154.1, 154.0, 53.1, 53.0, 128.1, 128.0, 54.1, 54.0, 129.1, 129.0, 130.1, 130.0, 131.1, 131.0, 132.1, 132.0, 133.1, 133.0, 147.1, 147.0, 134.1, 134.0, 135.1, 135.0, 136.1, 136.0, 55.1, 55.0, 137.1, 137.0, 56.1, 56.0, 138.1, 138.0, 57.1, 57.0, 155.1, 155.0, 58.1, 58.0, 156.1, 156.0, 59.1, 59.0, 157.1, 157.0, 158.1, 158.0, 159.1, 159.0, 160.1, 160.0, 161.1, 161.0, 162.1, 162.0, 163.1, 163.0, 164.1, 164.0, 165.1, 165.0, 60.1, 60.0, 61.1, 61.0, 166.1, 166.0, 166.0, 139.1, 139.0, 167.1, 167.0, 167.0, 148.1, 148.0, 168.1, 168.0, 168.0, 149.1, 149.0, 150.1, 150.0, 169.1, 169.0, 169.0, 62.1, 62.0, 166.1, 166.0, 170.1, 170.0, 170.0, 167.1, 167.0, 171.1, 171.0, 171.0, 168.1, 168.0, 172.1, 172.0, 172.0, 169.1, 169.0, 151.1, 151.0, 106.1, 106.0, 107.1, 107.0, 108.1, 108.0, 152.1, 152.0, 170.1, 170.0, 63.1, 63.0, 109.1, 109.0, 64.1, 64.0, 110.1, 110.0, 65.1, 65.0, 111.1, 111.0, 66.1, 66.0, 163.1, 163.0, 67.1, 67.0, 164.1, 164.0, 165.1, 165.0, 171.1, 171.0, 112.1, 112.0, 153.1, 153.0, 113.1, 113.0, 154.1, 154.0, 114.1, 114.0, 172.1, 172.0, 166.1, 166.0, 173.1, 173.0, 173.0, 68.1, 68.0, 167.1, 167.0, 69.1, 69.0, 168.1, 168.0, 70.1, 70.0, 71.1, 71.0, 169.1, 169.0, 173.1, 173.0, 115.1, 115.0, 174.1, 174.0, 174.0, 170.1, 170.0, 175.1, 175.0, 175.0, 140.1, 140.0, 155.1, 155.0, 156.1, 156.0, 176.1, 176.0, 176.0, 141.1, 141.0, 142.1, 142.0, 157.1, 157.0, 158.1, 158.0, 159.1, 159.0, 160.1, 160.0, 116.1, 116.0, 161.1, 161.0, 117.1, 117.0, 162.1, 162.0, 171.1, 171.0, 172.1, 172.0, 174.1, 174.0, 118.1, 118.0, 175.1, 175.0, 176.1, 176.0, 177.1, 177.0, 177.0, 143.1, 143.0, 72.1, 72.0, 178.1, 178.0, 178.0, 179.1, 179.0, 179.0, 73.1, 73.0, 173.1, 173.0, 119.1, 119.0, 177.1, 177.0, 180.1, 180.0, 180.0, 144.1, 144.0, 181.1, 181.0, 181.0, 178.1, 178.0, 179.1, 179.0, 182.1, 182.0, 182.0, 145.1, 145.0, 146.1, 146.0, 120.1, 120.0, 180.1, 180.0, 181.1, 181.0, 183.1, 183.0, 183.0, 121.1, 121.0, 122.1, 122.0, 74.1, 74.0, 123.1, 123.0, 75.1, 75.0, 124.1, 124.0, 125.1, 125.0, 147.1, 147.0, 76.1, 76.0, 163.1, 163.0, 77.1, 77.0, 164.1, 164.0, 78.1, 78.0, 165.1, 165.0, 79.1, 79.0, 166.1, 166.0, 167.1, 167.0, 174.1, 174.0, 168.1, 168.0, 175.1, 175.0, 80.1, 80.0, 176.1, 176.0, 169.1, 169.0, 177.1, 177.0, 178.1, 178.0, 179.1, 179.0, 182.1, 182.0, 170.1, 170.0, 180.1, 180.0, 181.1, 181.0, 81.1, 81.0, 183.1, 183.0, 82.1, 82.0, 184.1, 184.0, 184.0, 185.1, 185.0, 185.0, 148.1, 148.0, 186.1, 186.0, 186.0, 83.1, 83.0, 149.1, 149.0, 150.1, 150.0, 84.1, 84.0, 184.1, 184.0, 185.1, 185.0, 171.1, 171.0, 172.1, 172.0, 186.1, 186.0, 187.1, 187.0, 187.0, 151.1, 151.0, 188.1, 188.0, 188.0, 189.1, 189.0, 189.0, 152.1, 152.0, 190.1, 190.0, 190.0, 173.1, 173.0, 187.1, 187.0, 182.1, 182.0, 191.1, 191.0, 191.0, 85.1, 85.0, 188.1, 188.0, 189.1, 189.0, 86.1, 86.0, 87.1, 87.0, 126.1, 126.0, 88.1, 88.0, 127.1, 127.0, 89.1, 89.0, 183.1, 183.0, 90.1, 90.0, 153.1, 153.0, 154.1, 154.0, 190.1, 190.0, 191.1, 191.0, 184.1, 184.0, 185.1, 185.0, 128.1, 128.0, 91.1, 91.0, 92.1, 92.0, 93.1, 93.0, 129.1, 129.0, 94.1, 94.0, 130.1, 130.0, 95.1, 95.0, 131.1, 131.0, 132.1, 132.0, 96.1, 96.0, 133.1, 133.0, 134.1, 134.0, 135.1, 135.0, 136.1, 136.0, 137.1, 137.0, 138.1, 138.0, 155.1, 155.0, 156.1, 156.0, 157.1, 157.0, 158.1, 158.0, 159.1, 159.0, 160.1, 160.0, 97.1, 97.0, 161.1, 161.0, 162.1, 162.0, 174.1, 174.0, 98.1, 98.0, 175.1, 175.0, 176.1, 176.0, 139.1, 139.0, 0.1, 0.0, 177.1, 177.0, 1.1, 1.0, 178.1, 178.0, 2.1, 2.0, 179.1, 179.0, 3.1, 3.0, 180.1, 180.0, 4.1, 4.0, 5.1, 5.0, 99.1, 99.0, 6.1, 6.0, 100.1, 100.0, 7.1, 7.0, 181.1, 181.0, 8.1, 8.0, 186.1, 186.0, 9.1, 9.0, 187.1, 187.0, 188.1, 188.0, 10.1, 10.0, 182.1, 182.0, 11.1, 11.0, 189.1, 189.0, 12.1, 12.0, 190.1, 190.0, 13.1, 13.0, 191.1, 191.0, 192.1, 192.0, 192.0, 14.1, 14.0, 193.1, 193.0, 193.0, 15.1, 15.0, 183.1, 183.0, 16.1, 16.0, 194.1, 194.0, 194.0, 17.1, 17.0, 163.1, 163.0, 18.1, 18.0, 164.1, 164.0, 19.1, 19.0, 165.1, 165.0, 20.1, 20.0, 166.1, 166.0, 21.1, 21.0, 167.1, 167.0, 22.1, 22.0, 168.1, 168.0, 169.1, 169.0, 184.1, 184.0, 185.1, 185.0, 101.1, 101.0, 102.1, 102.0, 170.1, 170.0, 192.1, 192.0, 103.1, 103.0, 193.1, 193.0, 23.1, 23.0, 104.1, 104.0, 24.1, 24.0, 105.1, 105.0, 25.1, 25.0, 140.1, 140.0, 26.1, 26.0, 27.1, 27.0, 141.1, 141.0, 142.1, 142.0, 194.1, 194.0, 171.1, 171.0, 172.1, 172.0, 195.1, 195.0, 195.0, 196.1, 196.0, 196.0, 28.1, 28.0, 197.1, 197.0, 197.0, 29.1, 29.0, 198.1, 198.0, 198.0, 30.1, 30.0, 199.1, 199.0, 199.0, 31.1, 31.0, 32.1, 32.0, 143.1, 143.0, 173.1, 173.0, 195.1, 195.0, 196.1, 196.0, 192.1, 192.0, 197.1, 197.0, 193.1, 193.0, 198.1, 198.0, 144.1, 144.0, 33.1, 33.0, 186.1, 186.0, 34.1, 34.0, 187.1, 187.0, 35.1, 35.0, 145.1, 145.0, 36.1, 36.0, 37.1, 37.0, 38.1, 38.0, 146.1, 146.0, 188.1, 188.0, 189.1, 189.0, 190.1, 190.0, 191.1, 191.0, 194.1, 194.0, 199.1, 199.0, 200.1, 200.0, 200.0, 39.1, 39.0, 147.1, 147.0, 40.1, 40.0, 41.1, 41.0, 42.1, 42.0, 106.1, 106.0, 43.1, 43.0, 107.1, 107.0, 44.1, 44.0, 108.1, 108.0, 109.1, 109.0, 195.1, 195.0, 110.1, 110.0, 196.1, 196.0, 111.1, 111.0, 197.1, 197.0, 198.1, 198.0, 174.1, 174.0, 200.1, 200.0, 175.1, 175.0, 176.1, 176.0, 112.1, 112.0, 113.1, 113.0, 177.1, 177.0, 114.1, 114.0, 148.1, 148.0, 178.1, 178.0, 179.1, 179.0, 149.1, 149.0, 150.1, 150.0, 45.1, 45.0, 180.1, 180.0, 199.1, 199.0, 181.1, 181.0, 115.1, 115.0, 151.1, 151.0, 182.1, 182.0, 152.1, 152.0, 46.1, 46.0, 47.1, 47.0, 192.1, 192.0, 48.1, 48.0, 193.1, 193.0, 183.1, 183.0, 200.1, 200.0, 116.1, 116.0, 117.1, 117.0, 49.1, 49.0, 50.1, 50.0, 118.1, 118.0, 153.1, 153.0, 51.1, 51.0, 154.1, 154.0, 52.1, 52.0, 53.1, 53.0, 184.1, 184.0, 54.1, 54.0, 185.1, 185.0, 194.1, 194.0, 119.1, 119.0, 55.1, 55.0, 56.1, 56.0, 120.1, 120.0, 57.1, 57.0, 195.1, 195.0, 58.1, 58.0, 155.1, 155.0, 121.1, 121.0, 59.1, 59.0, 122.1, 122.0, 123.1, 123.0, 124.1, 124.0, 125.1, 125.0, 156.1, 156.0, 157.1, 157.0, 158.1, 158.0, 159.1, 159.0, 160.1, 160.0, 60.1, 60.0, 61.1, 61.0, 161.1, 161.0, 162.1, 162.0, 196.1, 196.0, 197.1, 197.0, 198.1, 198.0, 199.1, 199.0, 62.1, 62.0, 186.1, 186.0, 187.1, 187.0, 188.1, 188.0, 189.1, 189.0, 190.1, 190.0, 191.1, 191.0, 63.1, 63.0, 200.1, 200.0, 64.1, 64.0, 65.1, 65.0, 66.1, 66.0, 163.1, 163.0, 67.1, 67.0, 164.1, 164.0, 165.1, 165.0, 166.1, 166.0, 167.1, 167.0, 168.1, 168.0, 169.1, 169.0, 68.1, 68.0, 69.1, 69.0, 70.1, 70.0, 71.1, 71.0, 126.1, 126.0, 170.1, 170.0, 127.1, 127.0, 128.1, 128.0, 171.1, 171.0, 172.1, 172.0, 129.1, 129.0, 130.1, 130.0, 131.1, 131.0, 132.1, 132.0, 133.1, 133.0, 134.1, 134.0, 135.1, 135.0, 136.1, 136.0, 137.1, 137.0, 138.1, 138.0, 72.1, 72.0, 173.1, 173.0, 192.1, 192.0, 73.1, 73.0, 193.1, 193.0, 139.1, 139.0, 74.1, 74.0, 194.1, 194.0, 75.1, 75.0, 76.1, 76.0, 77.1, 77.0, 78.1, 78.0, 79.1, 79.0, 195.1, 195.0, 80.1, 80.0, 174.1, 174.0, 175.1, 175.0, 81.1, 81.0, 176.1, 176.0, 82.1, 82.0, 177.1, 177.0, 178.1, 178.0, 179.1, 179.0, 196.1, 196.0, 83.1, 83.0, 197.1, 197.0, 198.1, 198.0, 84.1, 84.0, 180.1, 180.0, 181.1, 181.0, 199.1, 199.0, 182.1, 182.0, 140.1, 140.0, 141.1, 141.0, 142.1, 142.0, 85.1, 85.0, 183.1, 183.0, 200.1, 200.0, 86.1, 86.0, 87.1, 87.0, 88.1, 88.0, 89.1, 89.0, 90.1, 90.0, 143.1, 143.0, 184.1, 184.0, 91.1, 91.0, 92.1, 92.0, 93.1, 93.0, 185.1, 185.0, 94.1, 94.0, 144.1, 144.0, 95.1, 95.0, 96.1, 96.0, 145.1, 145.0, 146.1, 146.0, 97.1, 97.0, 147.1, 147.0, 98.1, 98.0, 99.1, 99.0, 100.1, 100.0, 186.1, 186.0, 187.1, 187.0, 188.1, 188.0, 148.1, 148.0, 189.1, 189.0, 190.1, 190.0, 149.1, 149.0, 150.1, 150.0, 191.1, 191.0, 151.1, 151.0, 152.1, 152.0, 101.1, 101.0, 102.1, 102.0, 103.1, 103.0, 104.1, 104.0, 105.1, 105.0, 153.1, 153.0, 154.1, 154.0, 155.1, 155.0, 192.1, 192.0, 193.1, 193.0, 156.1, 156.0, 157.1, 157.0, 158.1, 158.0, 159.1, 159.0, 160.1, 160.0, 161.1, 161.0, 162.1, 162.0, 194.1, 194.0, 106.1, 106.0, 107.1, 107.0, 108.1, 108.0, 109.1, 109.0, 195.1, 195.0, 110.1, 110.0, 111.1, 111.0, 163.1, 163.0, 0.1, 0.0, 112.1, 112.0, 1.1, 1.0, 113.1, 113.0, 2.1, 2.0, 114.1, 114.0, 3.1, 3.0, 164.1, 164.0, 4.1, 4.0, 5.1, 5.0, 165.1, 165.0, 6.1, 6.0, 166.1, 166.0, 7.1, 7.0, 167.1, 167.0, 8.1, 8.0, 168.1, 168.0, 9.1, 9.0, 115.1, 115.0, 169.1, 169.0, 10.1, 10.0, 170.1, 170.0, 11.1, 11.0, 196.1, 196.0, 12.1, 12.0, 197.1, 197.0, 13.1, 13.0, 198.1, 198.0, 199.1, 199.0, 14.1, 14.0, 171.1, 171.0, 15.1, 15.0, 172.1, 172.0, 16.1, 16.0, 200.1, 200.0, 17.1, 17.0, 116.1, 116.0, 18.1, 18.0, 117.1, 117.0, 19.1, 19.0, 20.1, 20.0, 21.1, 21.0, 118.1, 118.0, 22.1, 22.0, 173.1, 173.0, 119.1, 119.0, 23.1, 23.0, 24.1, 24.0, 25.1, 25.0, 120.1, 120.0, 26.1, 26.0, 27.1, 27.0, 121.1, 121.0, 122.1, 122.0, 123.1, 123.0, 124.1, 124.0, 125.1, 125.0, 28.1, 28.0, 29.1, 29.0, 30.1, 30.0, 31.1, 31.0, 32.1, 32.0, 174.1, 174.0, 33.1, 33.0, 175.1, 175.0, 34.1, 34.0, 176.1, 176.0, 35.1, 35.0, 177.1, 177.0, 36.1, 36.0, 37.1, 37.0, 38.1, 38.0, 178.1, 178.0, 179.1, 179.0, 180.1, 180.0, 181.1, 181.0, 39.1, 39.0, 182.1, 182.0, 40.1, 40.0, 41.1, 41.0, 42.1, 42.0, 43.1, 43.0, 44.1, 44.0, 183.1, 183.0, 126.1, 126.0, 127.1, 127.0, 184.1, 184.0, 128.1, 128.0, 45.1, 45.0, 185.1, 185.0, 129.1, 129.0, 130.1, 130.0, 131.1, 131.0, 132.1, 132.0, 133.1, 133.0, 134.1, 134.0, 135.1, 135.0, 136.1, 136.0, 46.1, 46.0, 137.1, 137.0, 47.1, 47.0, 138.1, 138.0, 48.1, 48.0, 49.1, 49.0, 50.1, 50.0, 139.1, 139.0, 51.1, 51.0, 52.1, 52.0, 53.1, 53.0, 54.1, 54.0, 186.1, 186.0, 187.1, 187.0, 55.1, 55.0, 188.1, 188.0, 56.1, 56.0, 189.1, 189.0, 57.1, 57.0, 190.1, 190.0, 58.1, 58.0, 191.1, 191.0, 59.1, 59.0, 60.1, 60.0, 61.1, 61.0, 62.1, 62.0, 140.1, 140.0, 63.1, 63.0, 141.1, 141.0, 64.1, 64.0, 142.1, 142.0, 65.1, 65.0, 66.1, 66.0, 67.1, 67.0, 143.1, 143.0, 192.1, 192.0, 68.1, 68.0, 193.1, 193.0, 69.1, 69.0, 70.1, 70.0, 71.1, 71.0, 144.1, 144.0, 145.1, 145.0, 146.1, 146.0, 194.1, 194.0, 147.1, 147.0, 72.1, 72.0, 73.1, 73.0, 148.1, 148.0, 149.1, 149.0, 74.1, 74.0, 150.1, 150.0, 75.1, 75.0, 76.1, 76.0, 77.1, 77.0, 151.1, 151.0, 78.1, 78.0, 79.1, 79.0, 152.1, 152.0, 80.1, 80.0, 81.1, 81.0, 82.1, 82.0, 83.1, 83.0, 84.1, 84.0, 85.1, 85.0, 86.1, 86.0, 87.1, 87.0, 88.1, 88.0, 89.1, 89.0, 90.1, 90.0, 91.1, 91.0, 92.1, 92.0, 93.1, 93.0, 94.1, 94.0, 95.1, 95.0, 96.1, 96.0, 97.1, 97.0, 98.1, 98.0}
-
-// Advanced multi = 40
-var Advanced = []float32{
-	0.1, 0.0, 0.0, 1.1, 1.0, 1.0, 2.1, 2.0, 2.0, 3.1, 3.0, 3.0, 4.1, 4.0, 4.0, 5.1, 5.0, 5.0, 6.1, 6.0, 6.0, 7.1, 7.0, 7.0, 8.1, 8.0, 8.0, 9.1, 9.0, 9.0, 10.1, 10.0, 10.0, 11.1, 11.0, 11.0, 12.1, 12.0, 12.0, 13.1, 13.0, 13.0, 0.1, 0.0, 14.1, 14.0, 14.0, 1.1, 1.0, 15.1, 15.0, 15.0, 2.1, 2.0, 16.1, 16.0, 16.0, 3.1, 3.0, 17.1, 17.0, 17.0, 4.1, 4.0, 18.1, 18.0, 18.0, 5.1, 5.0, 19.1, 19.0, 19.0, 6.1, 6.0, 20.1, 20.0, 20.0, 7.1, 7.0, 21.1, 21.0, 21.0, 8.1, 8.0, 22.1, 22.0, 22.0, 9.1, 9.0, 23.1, 23.0, 23.0, 10.1, 10.0, 24.1, 24.0, 24.0, 11.1, 11.0, 25.1, 25.0, 25.0, 12.1, 12.0, 26.1, 26.0, 26.0, 13.1, 13.0, 27.1, 27.0, 27.0, 14.1, 14.0, 28.1, 28.0, 28.0, 15.1, 15.0, 29.1, 29.0, 29.0, 16.1, 16.0, 30.1, 30.0, 30.0, 17.1, 17.0, 31.1, 31.0, 31.0, 18.1, 18.0, 32.1, 32.0, 32.0, 19.1, 19.0, 33.1, 33.0, 33.0, 20.1, 20.0, 34.1, 34.0, 34.0, 21.1, 21.0, 35.1, 35.0, 35.0, 22.1, 22.0, 36.1, 36.0, 36.0, 23.1, 23.0, 37.1, 37.0, 37.0, 24.1, 24.0, 38.1, 38.0, 38.0, 25.1, 25.0, 39.1, 39.0, 39.0, 26.1, 26.0, 40.1, 40.0, 40.0, 27.1, 27.0, 41.1, 41.0, 41.0, 28.1, 28.0, 42.1, 42.0, 42.0, 29.1, 29.0, 43.1, 43.0, 43.0, 30.1, 30.0, 44.1, 44.0, 44.0, 31.1, 31.0, 45.1, 45.0, 45.0, 32.1, 32.0, 46.1, 46.0, 46.0, 33.1, 33.0, 47.1, 47.0, 47.0, 34.1, 34.0, 48.1, 48.0, 48.0, 35.1, 35.0, 49.1, 49.0, 49.0, 36.1, 36.0, 50.1, 50.0, 50.0, 37.1, 37.0, 51.1, 51.0, 51.0, 38.1, 38.0, 52.1, 52.0, 52.0, 39.1, 39.0, 53.1, 53.0, 53.0, 40.1, 40.0, 54.1, 54.0, 54.0, 41.1, 41.0, 55.1, 55.0, 55.0, 42.1, 42.0, 56.1, 56.0, 56.0, 43.1, 43.0, 57.1, 57.0, 57.0, 44.1, 44.0, 58.1, 58.0, 58.0, 45.1, 45.0, 59.1, 59.0, 59.0, 46.1, 46.0, 60.1, 60.0, 60.0, 47.1, 47.0, 61.1, 61.0, 61.0, 48.1, 48.0, 62.1, 62.0, 62.0, 49.1, 49.0, 63.1, 63.0, 63.0, 50.1, 50.0, 64.1, 64.0, 64.0, 51.1, 51.0, 65.1, 65.0, 65.0, 52.1, 52.0, 66.1, 66.0, 66.0, 53.1, 53.0, 0.1, 0.0, 54.1, 54.0, 1.1, 1.0, 55.1, 55.0, 2.1, 2.0, 56.1, 56.0, 3.1, 3.0, 57.1, 57.0, 4.1, 4.0, 58.1, 58.0, 5.1, 5.0, 59.1, 59.0, 6.1, 6.0, 60.1, 60.0, 7.1, 7.0, 61.1, 61.0, 8.1, 8.0, 62.1, 62.0, 9.1, 9.0, 63.1, 63.0, 10.1, 10.0, 64.1, 64.0, 11.1, 11.0, 65.1, 65.0, 12.1, 12.0, 66.1, 66.0, 13.1, 13.0, 67.1, 67.0, 67.0, 14.1, 14.0, 68.1, 68.0, 68.0, 15.1, 15.0, 69.1, 69.0, 69.0, 16.1, 16.0, 70.1, 70.0, 70.0, 17.1, 17.0, 71.1, 71.0, 71.0, 18.1, 18.0, 72.1, 72.0, 72.0, 19.1, 19.0, 73.1, 73.0, 73.0, 20.1, 20.0, 74.1, 74.0, 74.0, 21.1, 21.0, 75.1, 75.0, 75.0, 22.1, 22.0, 76.1, 76.0, 76.0, 23.1, 23.0, 77.1, 77.0, 77.0, 24.1, 24.0, 78.1, 78.0, 78.0, 25.1, 25.0, 79.1, 79.0, 79.0, 26.1, 26.0, 80.1, 80.0, 80.0, 27.1, 27.0, 67.1, 67.0, 28.1, 28.0, 68.1, 68.0, 29.1, 29.0, 69.1, 69.0, 30.1, 30.0, 70.1, 70.0, 31.1, 31.0, 71.1, 71.0, 32.1, 32.0, 72.1, 72.0, 33.1, 33.0, 73.1, 73.0, 34.1, 34.0, 74.1, 74.0, 35.1, 35.0, 75.1, 75.0, 36.1, 36.0, 76.1, 76.0, 37.1, 37.0, 77.1, 77.0, 38.1, 38.0, 78.1, 78.0, 39.1, 39.0, 79.1, 79.0, 40.1, 40.0, 80.1, 80.0, 41.1, 41.0, 81.1, 81.0, 81.0, 42.1, 42.0, 82.1, 82.0, 82.0, 43.1, 43.0, 83.1, 83.0, 83.0, 44.1, 44.0, 84.1, 84.0, 84.0, 45.1, 45.0, 85.1, 85.0, 85.0, 46.1, 46.0, 86.1, 86.0, 86.0, 47.1, 47.0, 87.1, 87.0, 87.0, 48.1, 48.0, 88.1, 88.0, 88.0, 49.1, 49.0, 89.1, 89.0, 89.0, 50.1, 50.0, 90.1, 90.0, 90.0, 51.1, 51.0, 91.1, 91.0, 91.0, 52.1, 52.0, 53.1, 53.0, 92.1, 92.0, 92.0, 54.1, 54.0, 93.1, 93.0, 93.0, 55.1, 55.0, 81.1, 81.0, 56.1, 56.0, 82.1, 82.0, 57.1, 57.0, 83.1, 83.0, 58.1, 58.0, 84.1, 84.0, 59.1, 59.0, 85.1, 85.0, 60.1, 60.0, 86.1, 86.0, 61.1, 61.0, 87.1, 87.0, 62.1, 62.0, 88.1, 88.0, 63.1, 63.0, 89.1, 89.0, 64.1, 64.0, 90.1, 90.0, 65.1, 65.0, 91.1, 91.0, 66.1, 66.0, 92.1, 92.0, 93.1, 93.0, 94.1, 94.0, 94.0, 95.1, 95.0, 95.0, 96.1, 96.0, 96.0, 97.1, 97.0, 97.0, 98.1, 98.0, 98.0, 99.1, 99.0, 99.0, 100.1, 100.0, 100.0, 101.1, 101.0, 101.0, 102.1, 102.0, 102.0, 103.1, 103.0, 103.0, 104.1, 104.0, 104.0, 105.1, 105.0, 105.0, 106.1, 106.0, 106.0, 67.1, 67.0, 107.1, 107.0, 107.0, 68.1, 68.0, 94.1, 94.0, 69.1, 69.0, 95.1, 95.0, 70.1, 70.0, 96.1, 96.0, 71.1, 71.0, 97.1, 97.0, 72.1, 72.0, 98.1, 98.0, 73.1, 73.0, 99.1, 99.0, 74.1, 74.0, 100.1, 100.0, 75.1, 75.0, 101.1, 101.0, 76.1, 76.0, 102.1, 102.0, 77.1, 77.0, 103.1, 103.0, 78.1, 78.0, 104.1, 104.0, 79.1, 79.0, 105.1, 105.0, 80.1, 80.0, 106.1, 106.0, 107.1, 107.0, 108.1, 108.0, 108.0, 109.1, 109.0, 109.0, 110.1, 110.0, 110.0, 111.1, 111.0, 111.0, 112.1, 112.0, 112.0, 113.1, 113.0, 113.0, 114.1, 114.0, 114.0, 115.1, 115.0, 115.0, 116.1, 116.0, 116.0, 117.1, 117.0, 117.0, 118.1, 118.0, 118.0, 119.1, 119.0, 119.0, 120.1, 120.0, 120.0, 81.1, 81.0, 121.1, 121.0, 121.0, 82.1, 82.0, 108.1, 108.0, 83.1, 83.0, 109.1, 109.0, 84.1, 84.0, 110.1, 110.0, 85.1, 85.0, 111.1, 111.0, 86.1, 86.0, 112.1, 112.0, 87.1, 87.0, 113.1, 113.0, 88.1, 88.0, 114.1, 114.0, 89.1, 89.0, 115.1, 115.0, 90.1, 90.0, 116.1, 116.0, 91.1, 91.0, 117.1, 117.0, 92.1, 92.0, 0.1, 0.0, 93.1, 93.0, 1.1, 1.0, 118.1, 118.0, 2.1, 2.0, 119.1, 119.0, 3.1, 3.0, 120.1, 120.0, 4.1, 4.0, 121.1, 121.0, 5.1, 5.0, 122.1, 122.0, 122.0, 6.1, 6.0, 123.1, 123.0, 123.0, 7.1, 7.0, 124.1, 124.0, 124.0, 8.1, 8.0, 125.1, 125.0, 125.0, 9.1, 9.0, 126.1, 126.0, 126.0, 10.1, 10.0, 127.1, 127.0, 127.0, 11.1, 11.0, 128.1, 128.0, 128.0, 12.1, 12.0, 129.1, 129.0, 129.0, 13.1, 13.0, 130.1, 130.0, 130.0, 14.1, 14.0, 94.1, 94.0, 15.1, 15.0, 95.1, 95.0, 16.1, 16.0, 96.1, 96.0, 17.1, 17.0, 97.1, 97.0, 18.1, 18.0, 98.1, 98.0, 19.1, 19.0, 99.1, 99.0, 20.1, 20.0, 100.1, 100.0, 21.1, 21.0, 101.1, 101.0, 22.1, 22.0, 102.1, 102.0, 23.1, 23.0, 103.1, 103.0, 24.1, 24.0, 104.1, 104.0, 25.1, 25.0, 105.1, 105.0, 26.1, 26.0, 106.1, 106.0, 27.1, 27.0, 107.1, 107.0, 28.1, 28.0, 122.1, 122.0, 29.1, 29.0, 123.1, 123.0, 30.1, 30.0, 124.1, 124.0, 31.1, 31.0, 125.1, 125.0, 32.1, 32.0, 126.1, 126.0, 33.1, 33.0, 127.1, 127.0, 34.1, 34.0, 128.1, 128.0, 35.1, 35.0, 129.1, 129.0, 36.1, 36.0, 130.1, 130.0, 37.1, 37.0, 108.1, 108.0, 38.1, 38.0, 109.1, 109.0, 39.1, 39.0, 110.1, 110.0, 40.1, 40.0, 111.1, 111.0, 41.1, 41.0, 112.1, 112.0, 42.1, 42.0, 113.1, 113.0, 43.1, 43.0, 114.1, 114.0, 44.1, 44.0, 115.1, 115.0, 45.1, 45.0, 116.1, 116.0, 46.1, 46.0, 117.1, 117.0, 47.1, 47.0, 118.1, 118.0, 48.1, 48.0, 119.1, 119.0, 49.1, 49.0, 120.1, 120.0, 50.1, 50.0, 121.1, 121.0, 51.1, 51.0, 131.1, 131.0, 131.0, 52.1, 52.0, 53.1, 53.0, 132.1, 132.0, 132.0, 54.1, 54.0, 133.1, 133.0, 133.0, 55.1, 55.0, 134.1, 134.0, 134.0, 56.1, 56.0, 135.1, 135.0, 135.0, 57.1, 57.0, 136.1, 136.0, 136.0, 58.1, 58.0, 137.1, 137.0, 137.0, 59.1, 59.0, 138.1, 138.0, 138.0, 60.1, 60.0, 139.1, 139.0, 139.0, 61.1, 61.0, 140.1, 140.0, 140.0, 62.1, 62.0, 141.1, 141.0, 141.0, 63.1, 63.0, 142.1, 142.0, 142.0, 64.1, 64.0, 143.1, 143.0, 143.0, 65.1, 65.0, 131.1, 131.0, 66.1, 66.0, 132.1, 132.0, 144.1, 144.0, 144.0, 133.1, 133.0, 145.1, 145.0, 145.0, 134.1, 134.0, 146.1, 146.0, 146.0, 135.1, 135.0, 147.1, 147.0, 147.0, 136.1, 136.0, 148.1, 148.0, 148.0, 137.1, 137.0, 149.1, 149.0, 149.0, 138.1, 138.0, 150.1, 150.0, 150.0, 139.1, 139.0, 151.1, 151.0, 151.0, 140.1, 140.0, 152.1, 152.0, 152.0, 141.1, 141.0, 153.1, 153.0, 153.0, 142.1, 142.0, 122.1, 122.0, 143.1, 143.0, 123.1, 123.0, 154.1, 154.0, 154.0, 124.1, 124.0, 155.1, 155.0, 155.0, 125.1, 125.0, 67.1, 67.0, 126.1, 126.0, 68.1, 68.0, 127.1, 127.0, 69.1, 69.0, 128.1, 128.0, 70.1, 70.0, 129.1, 129.0, 71.1, 71.0, 130.1, 130.0, 72.1, 72.0, 144.1, 144.0, 73.1, 73.0, 145.1, 145.0, 74.1, 74.0, 146.1, 146.0, 75.1, 75.0, 147.1, 147.0, 76.1, 76.0, 148.1, 148.0, 77.1, 77.0, 149.1, 149.0, 78.1, 78.0, 150.1, 150.0, 79.1, 79.0, 151.1, 151.0, 80.1, 80.0, 152.1, 152.0, 153.1, 153.0, 154.1, 154.0, 155.1, 155.0, 156.1, 156.0, 156.0, 157.1, 157.0, 157.0, 158.1, 158.0, 158.0, 159.1, 159.0, 159.0, 160.1, 160.0, 160.0, 161.1, 161.0, 161.0, 162.1, 162.0, 162.0, 163.1, 163.0, 163.0, 164.1, 164.0, 164.0, 165.1, 165.0, 165.0, 166.1, 166.0, 166.0, 167.1, 167.0, 167.0, 81.1, 81.0, 168.1, 168.0, 168.0, 82.1, 82.0, 169.1, 169.0, 169.0, 83.1, 83.0, 156.1, 156.0, 84.1, 84.0, 157.1, 157.0, 85.1, 85.0, 158.1, 158.0, 86.1, 86.0, 159.1, 159.0, 87.1, 87.0, 160.1, 160.0, 88.1, 88.0, 161.1, 161.0, 89.1, 89.0, 131.1, 131.0, 90.1, 90.0, 132.1, 132.0, 91.1, 91.0, 133.1, 133.0, 92.1, 92.0, 134.1, 134.0, 93.1, 93.0, 135.1, 135.0, 136.1, 136.0, 137.1, 137.0, 138.1, 138.0, 139.1, 139.0, 162.1, 162.0, 140.1, 140.0, 163.1, 163.0, 141.1, 141.0, 164.1, 164.0, 142.1, 142.0, 165.1, 165.0, 143.1, 143.0, 166.1, 166.0, 167.1, 167.0, 168.1, 168.0, 169.1, 169.0, 170.1, 170.0, 170.0, 171.1, 171.0, 171.0, 172.1, 172.0, 172.0, 173.1, 173.0, 173.0, 94.1, 94.0, 174.1, 174.0, 174.0, 95.1, 95.0, 175.1, 175.0, 175.0, 96.1, 96.0, 176.1, 176.0, 176.0, 97.1, 97.0, 144.1, 144.0, 98.1, 98.0, 145.1, 145.0, 99.1, 99.0, 146.1, 146.0, 100.1, 100.0, 147.1, 147.0, 101.1, 101.0, 148.1, 148.0, 102.1, 102.0, 149.1, 149.0, 103.1, 103.0, 150.1, 150.0, 104.1, 104.0, 151.1, 151.0, 105.1, 105.0, 152.1, 152.0, 106.1, 106.0, 153.1, 153.0, 107.1, 107.0, 154.1, 154.0, 155.1, 155.0, 170.1, 170.0, 171.1, 171.0, 172.1, 172.0, 173.1, 173.0, 174.1, 174.0, 175.1, 175.0, 176.1, 176.0, 177.1, 177.0, 177.0, 178.1, 178.0, 178.0, 179.1, 179.0, 179.0, 180.1, 180.0, 180.0, 108.1, 108.0, 181.1, 181.0, 181.0, 109.1, 109.0, 182.1, 182.0, 182.0, 110.1, 110.0, 156.1, 156.0, 111.1, 111.0, 157.1, 157.0, 112.1, 112.0, 158.1, 158.0, 113.1, 113.0, 159.1, 159.0, 114.1, 114.0, 160.1, 160.0, 115.1, 115.0, 161.1, 161.0, 116.1, 116.0, 177.1, 177.0, 117.1, 117.0, 178.1, 178.0, 118.1, 118.0, 179.1, 179.0, 119.1, 119.0, 180.1, 180.0, 120.1, 120.0, 181.1, 181.0, 121.1, 121.0, 182.1, 182.0, 183.1, 183.0, 183.0, 162.1, 162.0, 184.1, 184.0, 184.0, 163.1, 163.0, 185.1, 185.0, 185.0, 164.1, 164.0, 186.1, 186.0, 186.0, 165.1, 165.0, 187.1, 187.0, 187.0, 166.1, 166.0, 167.1, 167.0, 168.1, 168.0, 169.1, 169.0, 188.1, 188.0, 188.0, 189.1, 189.0, 189.0, 190.1, 190.0, 190.0, 191.1, 191.0, 191.0, 192.1, 192.0, 192.0, 193.1, 193.0, 193.0, 183.1, 183.0, 194.1, 194.0, 194.0, 184.1, 184.0, 195.1, 195.0, 195.0, 185.1, 185.0, 196.1, 196.0, 196.0, 186.1, 186.0, 197.1, 197.0, 197.0, 187.1, 187.0, 198.1, 198.0, 198.0, 199.1, 199.0, 199.0, 200.1, 200.0, 200.0, 188.1, 188.0, 189.1, 189.0, 190.1, 190.0, 191.1, 191.0, 122.1, 122.0, 192.1, 192.0, 123.1, 123.0, 193.1, 193.0, 124.1, 124.0, 194.1, 194.0, 125.1, 125.0, 170.1, 170.0, 126.1, 126.0, 171.1, 171.0, 127.1, 127.0, 172.1, 172.0, 128.1, 128.0, 173.1, 173.0, 129.1, 129.0, 174.1, 174.0, 130.1, 130.0, 175.1, 175.0, 176.1, 176.0, 195.1, 195.0, 196.1, 196.0, 197.1, 197.0, 198.1, 198.0, 199.1, 199.0, 200.1, 200.0, 0.1, 0.0, 1.1, 1.0, 2.1, 2.0, 3.1, 3.0, 4.1, 4.0, 177.1, 177.0, 5.1, 5.0, 178.1, 178.0, 6.1, 6.0, 179.1, 179.0, 7.1, 7.0, 180.1, 180.0, 8.1, 8.0, 181.1, 181.0, 9.1, 9.0, 182.1, 182.0, 10.1, 10.0, 11.1, 11.0, 12.1, 12.0, 13.1, 13.0, 14.1, 14.0, 15.1, 15.0, 16.1, 16.0, 17.1, 17.0, 18.1, 18.0, 19.1, 19.0, 20.1, 20.0, 21.1, 21.0, 183.1, 183.0, 22.1, 22.0, 184.1, 184.0, 23.1, 23.0, 131.1, 131.0, 24.1, 24.0, 132.1, 132.0, 25.1, 25.0, 133.1, 133.0, 26.1, 26.0, 134.1, 134.0, 27.1, 27.0, 135.1, 135.0, 28.1, 28.0, 136.1, 136.0, 29.1, 29.0, 137.1, 137.0, 30.1, 30.0, 138.1, 138.0, 31.1, 31.0, 139.1, 139.0, 32.1, 32.0, 140.1, 140.0, 33.1, 33.0, 141.1, 141.0, 34.1, 34.0, 142.1, 142.0, 35.1, 35.0, 143.1, 143.0, 36.1, 36.0, 185.1, 185.0, 37.1, 37.0, 186.1, 186.0, 38.1, 38.0, 187.1, 187.0, 39.1, 39.0, 188.1, 188.0, 40.1, 40.0, 144.1, 144.0, 41.1, 41.0, 145.1, 145.0, 42.1, 42.0, 146.1, 146.0, 43.1, 43.0, 147.1, 147.0, 44.1, 44.0, 148.1, 148.0, 45.1, 45.0, 149.1, 149.0, 46.1, 46.0, 150.1, 150.0, 47.1, 47.0, 151.1, 151.0, 48.1, 48.0, 152.1, 152.0, 49.1, 49.0, 153.1, 153.0, 50.1, 50.0, 154.1, 154.0, 51.1, 51.0, 155.1, 155.0, 52.1, 52.0, 53.1, 53.0, 189.1, 189.0, 54.1, 54.0, 190.1, 190.0, 55.1, 55.0, 191.1, 191.0, 56.1, 56.0, 192.1, 192.0, 57.1, 57.0, 193.1, 193.0, 58.1, 58.0, 156.1, 156.0, 59.1, 59.0, 157.1, 157.0, 60.1, 60.0, 158.1, 158.0, 61.1, 61.0, 159.1, 159.0, 62.1, 62.0, 160.1, 160.0, 63.1, 63.0, 161.1, 161.0, 64.1, 64.0, 194.1, 194.0, 65.1, 65.0, 195.1, 195.0, 66.1, 66.0, 196.1, 196.0, 197.1, 197.0, 198.1, 198.0, 162.1, 162.0, 199.1, 199.0, 163.1, 163.0, 200.1, 200.0, 164.1, 164.0, 165.1, 165.0, 166.1, 166.0, 167.1, 167.0, 168.1, 168.0, 169.1, 169.0, 67.1, 67.0, 68.1, 68.0, 69.1, 69.0, 70.1, 70.0, 71.1, 71.0, 72.1, 72.0, 73.1, 73.0, 74.1, 74.0, 75.1, 75.0, 76.1, 76.0, 77.1, 77.0, 78.1, 78.0, 79.1, 79.0, 80.1, 80.0, 170.1, 170.0, 171.1, 171.0, 172.1, 172.0, 173.1, 173.0, 174.1, 174.0, 175.1, 175.0, 176.1, 176.0, 81.1, 81.0, 82.1, 82.0, 83.1, 83.0, 84.1, 84.0, 85.1, 85.0, 86.1, 86.0, 177.1, 177.0, 87.1, 87.0, 178.1, 178.0, 88.1, 88.0, 179.1, 179.0, 89.1, 89.0, 180.1, 180.0, 90.1, 90.0, 181.1, 181.0, 91.1, 91.0, 182.1, 182.0, 92.1, 92.0, 93.1, 93.0, 183.1, 183.0, 184.1, 184.0, 94.1, 94.0, 95.1, 95.0, 96.1, 96.0, 97.1, 97.0, 98.1, 98.0, 99.1, 99.0, 100.1, 100.0, 101.1, 101.0, 102.1, 102.0, 103.1, 103.0, 104.1, 104.0, 105.1, 105.0, 106.1, 106.0, 107.1, 107.0, 185.1, 185.0, 186.1, 186.0, 187.1, 187.0, 188.1, 188.0, 108.1, 108.0, 109.1, 109.0, 110.1, 110.0, 111.1, 111.0, 112.1, 112.0, 113.1, 113.0, 114.1, 114.0, 115.1, 115.0, 116.1, 116.0, 117.1, 117.0, 118.1, 118.0, 119.1, 119.0, 189.1, 189.0, 120.1, 120.0, 190.1, 190.0, 121.1, 121.0, 191.1, 191.0, 192.1, 192.0, 193.1, 193.0, 194.1, 194.0, 195.1, 195.0, 196.1, 196.0, 197.1, 197.0, 198.1, 198.0, 199.1, 199.0, 200.1, 200.0, 122.1, 122.0, 123.1, 123.0, 124.1, 124.0, 125.1, 125.0, 126.1, 126.0, 127.1, 127.0, 128.1, 128.0, 129.1, 129.0, 130.1, 130.0, 131.1, 131.0, 132.1, 132.0, 133.1, 133.0, 134.1, 134.0, 135.1, 135.0, 136.1, 136.0, 137.1, 137.0, 138.1, 138.0, 139.1, 139.0, 140.1, 140.0, 141.1, 141.0, 142.1, 142.0, 143.1, 143.0, 144.1, 144.0, 145.1, 145.0, 146.1, 146.0, 147.1, 147.0, 148.1, 148.0, 149.1, 149.0, 150.1, 150.0, 151.1, 151.0, 152.1, 152.0, 153.1, 153.0, 154.1, 154.0, 155.1, 155.0, 156.1, 156.0, 157.1, 157.0, 158.1, 158.0, 159.1, 159.0, 160.1, 160.0, 161.1, 161.0, 162.1, 162.0, 163.1, 163.0, 164.1, 164.0, 165.1, 165.0, 166.1, 166.0, 167.1, 167.0, 168.1, 168.0, 169.1, 169.0, 170.1, 170.0, 171.1, 171.0, 172.1, 172.0, 173.1, 173.0, 174.1, 174.0, 175.1, 175.0, 176.1, 176.0, 177.1, 177.0, 178.1, 178.0, 179.1, 179.0, 180.1, 180.0, 181.1, 181.0, 182.1, 182.0, 0.1, 0.0, 1.1, 1.0, 2.1, 2.0, 3.1, 3.0, 4.1, 4.0, 5.1, 5.0, 6.1, 6.0, 7.1, 7.0, 8.1, 8.0, 9.1, 9.0, 183.1, 183.0, 10.1, 10.0, 184.1, 184.0, 11.1, 11.0, 12.1, 12.0, 13.1, 13.0, 14.1, 14.0, 15.1, 15.0, 16.1, 16.0, 17.1, 17.0, 18.1, 18.0, 19.1, 19.0, 20.1, 20.0, 21.1, 21.0, 22.1, 22.0, 23.1, 23.0, 24.1, 24.0, 25.1, 25.0, 26.1, 26.0, 27.1, 27.0, 185.1, 185.0, 28.1, 28.0, 186.1, 186.0, 29.1, 29.0, 187.1, 187.0, 30.1, 30.0, 188.1, 188.0, 31.1, 31.0, 32.1, 32.0, 33.1, 33.0, 34.1, 34.0, 35.1, 35.0, 36.1, 36.0, 37.1, 37.0, 38.1, 38.0, 39.1, 39.0, 40.1, 40.0, 41.1, 41.0, 42.1, 42.0, 43.1, 43.0, 44.1, 44.0, 45.1, 45.0, 46.1, 46.0, 189.1, 189.0, 47.1, 47.0, 190.1, 190.0, 48.1, 48.0, 191.1, 191.0, 49.1, 49.0, 192.1, 192.0, 50.1, 50.0, 193.1, 193.0, 51.1, 51.0, 52.1, 52.0, 53.1, 53.0, 54.1, 54.0, 55.1, 55.0, 56.1, 56.0, 57.1, 57.0, 58.1, 58.0, 194.1, 194.0, 59.1, 59.0, 195.1, 195.0, 60.1, 60.0, 196.1, 196.0, 61.1, 61.0, 197.1, 197.0, 62.1, 62.0, 198.1, 198.0, 63.1, 63.0, 199.1, 199.0, 64.1, 64.0, 200.1, 200.0, 65.1, 65.0, 66.1, 66.0, 67.1, 67.0, 68.1, 68.0, 69.1, 69.0, 70.1, 70.0, 71.1, 71.0, 72.1, 72.0, 73.1, 73.0, 74.1, 74.0, 75.1, 75.0, 76.1, 76.0, 77.1, 77.0, 78.1, 78.0, 79.1, 79.0, 80.1, 80.0, 81.1, 81.0, 82.1, 82.0, 83.1, 83.0, 84.1, 84.0, 85.1, 85.0, 86.1, 86.0, 87.1, 87.0, 88.1, 88.0, 89.1, 89.0, 90.1, 90.0, 91.1, 91.0, 92.1, 92.0, 93.1, 93.0, 94.1, 94.0, 95.1, 95.0, 96.1, 96.0, 97.1, 97.0, 98.1, 98.0, 99.1, 99.0, 100.1, 100.0, 101.1, 101.0, 102.1, 102.0, 103.1, 103.0, 104.1, 104.0, 105.1, 105.0, 106.1, 106.0, 107.1, 107.0, 108.1, 108.0, 109.1, 109.0, 110.1, 110.0, 111.1, 111.0, 112.1, 112.0, 113.1, 113.0, 114.1, 114.0, 115.1, 115.0, 116.1, 116.0, 117.1, 117.0, 118.1, 118.0, 119.1, 119.0, 120.1, 120.0, 121.1, 121.0, 122.1, 122.0, 123.1, 123.0, 124.1, 124.0, 125.1, 125.0, 126.1, 126.0, 127.1, 127.0, 128.1, 128.0, 129.1, 129.0, 130.1, 130.0, 131.1, 131.0, 132.1, 132.0, 133.1, 133.0, 134.1, 134.0, 135.1, 135.0, 136.1, 136.0, 137.1, 137.0, 138.1, 138.0, 139.1, 139.0, 140.1, 140.0, 141.1, 141.0, 142.1, 142.0, 143.1, 143.0, 144.1, 144.0, 145.1, 145.0, 146.1, 146.0, 147.1, 147.0, 148.1, 148.0, 149.1, 149.0, 150.1, 150.0, 151.1, 151.0, 152.1, 152.0, 153.1, 153.0, 154.1, 154.0, 155.1, 155.0, 156.1, 156.0, 157.1, 157.0, 158.1, 158.0, 159.1, 159.0, 160.1, 160.0, 161.1, 161.0, 162.1, 162.0, 163.1, 163.0, 164.1, 164.0, 165.1, 165.0, 166.1, 166.0, 167.1, 167.0, 168.1, 168.0, 169.1, 169.0, 170.1, 170.0, 171.1, 171.0, 172.1, 172.0, 173.1, 173.0, 174.1, 174.0, 175.1, 175.0, 176.1, 176.0, 177.1, 177.0, 178.1, 178.0, 179.1, 179.0, 180.1, 180.0, 181.1, 181.0, 182.1, 182.0, 183.1, 183.0, 184.1, 184.0, 185.1, 185.0, 186.1, 186.0, 187.1, 187.0, 188.1, 188.0, 189.1, 189.0, 190.1, 190.0, 191.1, 191.0, 192.1, 192.0, 193.1, 193.0, 194.1, 194.0, 195.1, 195.0, 196.1, 196.0, 197.1, 197.0, 198.1, 198.0, 199.1, 199.0, 200.1, 200.0, 0.1, 0.0, 1.1, 1.0, 2.1, 2.0, 3.1, 3.0, 4.1, 4.0, 5.1, 5.0, 6.1, 6.0, 7.1, 7.0, 8.1, 8.0, 9.1, 9.0, 10.1, 10.0, 11.1, 11.0, 12.1, 12.0, 13.1, 13.0, 14.1, 14.0, 15.1, 15.0, 16.1, 16.0, 17.1, 17.0, 18.1, 18.0, 19.1, 19.0, 20.1, 20.0, 21.1, 21.0, 22.1, 22.0, 23.1, 23.0, 24.1, 24.0, 25.1, 25.0, 26.1, 26.0, 27.1, 27.0, 28.1, 28.0, 29.1, 29.0, 30.1, 30.0, 31.1, 31.0, 32.1, 32.0, 33.1, 33.0, 34.1, 34.0, 35.1, 35.0, 36.1, 36.0, 37.1, 37.0, 38.1, 38.0, 39.1, 39.0, 40.1, 40.0, 41.1, 41.0, 42.1, 42.0, 43.1, 43.0, 44.1, 44.0, 45.1, 45.0, 46.1, 46.0, 47.1, 47.0, 48.1, 48.0, 49.1, 49.0, 50.1, 50.0, 51.1, 51.0, 52.1, 52.0, 53.1, 53.0, 54.1, 54.0, 55.1, 55.0, 56.1, 56.0, 57.1, 57.0, 58.1, 58.0, 59.1, 59.0, 60.1, 60.0, 61.1, 61.0, 62.1, 62.0, 63.1, 63.0, 64.1, 64.0, 65.1, 65.0, 66.1, 66.0}
-
-// Review each phrase is only repeated one time
-var Review = []float32{0.1, 0.0, 0.0, 1.1, 1.0, 1.0, 2.1, 2.0, 2.0, 3.1, 3.0, 3.0, 4.1, 4.0, 4.0, 5.1, 5.0, 5.0, 6.1, 6.0, 6.0, 7.1, 7.0, 7.0, 8.1, 8.0, 8.0, 9.1, 9.0, 9.0, 10.1, 10.0, 10.0, 11.1, 11.0, 11.0, 12.1, 12.0, 12.0, 13.1, 13.0, 13.0, 14.1, 14.0, 14.0, 15.1, 15.0, 15.0, 16.1, 16.0, 16.0, 17.1, 17.0, 17.0, 18.1, 18.0, 18.0, 19.1, 19.0, 19.0, 20.1, 20.0, 20.0, 21.1, 21.0, 21.0, 22.1, 22.0, 22.0, 23.1, 23.0, 23.0, 24.1, 24.0, 24.0, 25.1, 25.0, 25.0, 26.1, 26.0, 26.0, 27.1, 27.0, 27.0, 28.1, 28.0, 28.0, 29.1, 29.0, 29.0, 30.1, 30.0, 30.0, 31.1, 31.0, 31.0, 32.1, 32.0, 32.0, 33.1, 33.0, 33.0, 34.1, 34.0, 34.0, 35.1, 35.0, 35.0, 36.1, 36.0, 36.0, 37.1, 37.0, 37.0, 38.1, 38.0, 38.0, 39.1, 39.0, 39.0, 40.1, 40.0, 40.0, 41.1, 41.0, 41.0, 42.1, 42.0, 42.0, 43.1, 43.0, 43.0, 44.1, 44.0, 44.0, 45.1, 45.0, 45.0, 46.1, 46.0, 46.0, 47.1, 47.0, 47.0, 48.1, 48.0, 48.0, 49.1, 49.0, 49.0, 50.1, 50.0, 50.0, 51.1, 51.0, 51.0, 52.1, 52.0, 52.0, 53.1, 53.0, 53.0, 54.1, 54.0, 54.0, 55.1, 55.0, 55.0, 56.1, 56.0, 56.0, 57.1, 57.0, 57.0, 58.1, 58.0, 58.0, 59.1, 59.0, 59.0, 60.1, 60.0, 60.0, 61.1, 61.0, 61.0, 62.1, 62.0, 62.0, 63.1, 63.0, 63.0, 64.1, 64.0, 64.0, 65.1, 65.0, 65.0, 66.1, 66.0, 66.0, 67.1, 67.0, 67.0, 68.1, 68.0, 68.0, 69.1, 69.0, 69.0, 70.1, 70.0, 70.0, 71.1, 71.0, 71.0, 72.1, 72.0, 72.0, 73.1, 73.0, 73.0, 74.1, 74.0, 74.0, 75.1, 75.0, 75.0, 76.1, 76.0, 76.0, 77.1, 77.0, 77.0, 78.1, 78.0, 78.0, 79.1, 79.0, 79.0, 80.1, 80.0, 80.0, 81.1, 81.0, 81.0, 82.1, 82.0, 82.0, 83.1, 83.0, 83.0, 84.1, 84.0, 84.0, 85.1, 85.0, 85.0, 86.1, 86.0, 86.0, 87.1, 87.0, 87.0, 88.1, 88.0, 88.0, 89.1, 89.0, 89.0, 90.1, 90.0, 90.0, 91.1, 91.0, 91.0, 92.1, 92.0, 92.0, 93.1, 93.0, 93.0, 94.1, 94.0, 94.0, 95.1, 95.0, 95.0, 96.1, 96.0, 96.0, 97.1, 97.0, 97.0, 98.1, 98.0, 98.0, 99.1, 99.0, 99.0, 100.1, 100.0, 100.0, 101.1, 101.0, 101.0, 102.1, 102.0, 102.0, 103.1, 103.0, 103.0, 104.1, 104.0, 104.0, 105.1, 105.0, 105.0, 106.1, 106.0, 106.0, 107.1, 107.0, 107.0, 108.1, 108.0, 108.0, 109.1, 109.0, 109.0, 110.1, 110.0, 110.0, 111.1, 111.0, 111.0, 112.1, 112.0, 112.0, 113.1, 113.0, 113.0, 114.1, 114.0, 114.0, 115.1, 115.0, 115.0, 116.1, 116.0, 116.0, 117.1, 117.0, 117.0, 118.1, 118.0, 118.0, 119.1, 119.0, 119.0, 120.1, 120.0, 120.0, 121.1, 121.0, 121.0, 122.1, 122.0, 122.0, 123.1, 123.0, 123.0, 124.1, 124.0, 124.0, 125.1, 125.0, 125.0, 126.1, 126.0, 126.0, 127.1, 127.0, 127.0, 128.1, 128.0, 128.0, 129.1, 129.0, 129.0, 130.1, 130.0, 130.0, 131.1, 131.0, 131.0, 132.1, 132.0, 132.0, 133.1, 133.0, 133.0, 134.1, 134.0, 134.0, 135.1, 135.0, 135.0, 136.1, 136.0, 136.0, 137.1, 137.0, 137.0, 138.1, 138.0, 138.0, 139.1, 139.0, 139.0, 140.1, 140.0, 140.0, 141.1, 141.0, 141.0, 142.1, 142.0, 142.0, 143.1, 143.0, 143.0, 144.1, 144.0, 144.0, 145.1, 145.0, 145.0, 146.1, 146.0, 146.0, 147.1, 147.0, 147.0, 148.1, 148.0, 148.0, 149.1, 149.0, 149.0, 150.1, 150.0, 150.0, 151.1, 151.0, 151.0, 152.1, 152.0, 152.0, 153.1, 153.0, 153.0, 154.1, 154.0, 154.0, 155.1, 155.0, 155.0, 156.1, 156.0, 156.0, 157.1, 157.0, 157.0, 158.1, 158.0, 158.0, 159.1, 159.0, 159.0, 160.1, 160.0, 160.0, 161.1, 161.0, 161.0, 162.1, 162.0, 162.0, 163.1, 163.0, 163.0, 164.1, 164.0, 164.0, 165.1, 165.0, 165.0, 166.1, 166.0, 166.0, 167.1, 167.0, 167.0, 168.1, 168.0, 168.0, 169.1, 169.0, 169.0, 170.1, 170.0, 170.0, 171.1, 171.0, 171.0, 172.1, 172.0, 172.0, 173.1, 173.0, 173.0, 174.1, 174.0, 174.0, 175.1, 175.0, 175.0, 176.1, 176.0, 176.0, 177.1, 177.0, 177.0, 178.1, 178.0, 178.0, 179.1, 179.0, 179.0, 180.1, 180.0, 180.0, 181.1, 181.0, 181.0, 182.1, 182.0, 182.0, 183.1, 183.0, 183.0, 184.1, 184.0, 184.0, 185.1, 185.0, 185.0, 186.1, 186.0, 186.0, 187.1, 187.0, 187.0, 188.1, 188.0, 188.0, 189.1, 189.0, 189.0, 190.1, 190.0, 190.0, 191.1, 191.0, 191.0, 192.1, 192.0, 192.0, 193.1, 193.0, 193.0, 194.1, 194.0, 194.0, 195.1, 195.0, 195.0, 196.1, 196.0, 196.0, 197.1, 197.0, 197.0, 198.1, 198.0, 198.0, 199.1, 199.0, 199.0, 200.1, 200.0, 200.0}
+var (
+	Intermediate = []uint16{11, 10, 10, 21, 20, 20, 31, 30, 30, 41, 40, 40, 51, 50, 50, 61, 60, 60, 11, 10, 71, 70, 70, 21, 20, 81, 80, 80, 31, 30, 91, 90, 90, 41, 40, 51, 50, 101, 100, 100, 61, 60, 111, 110, 110, 71, 70, 121, 120, 120, 81, 80, 131, 130, 130, 91, 90, 141, 140, 140, 101, 100, 151, 150, 150, 111, 110, 161, 160, 160, 121, 120, 171, 170, 170, 131, 130, 181, 180, 180, 141, 140, 191, 190, 190, 151, 150, 201, 200, 200, 161, 160, 211, 210, 210, 171, 170, 221, 220, 220, 181, 180, 191, 190, 11, 10, 201, 200, 21, 20, 211, 210, 31, 30, 221, 220, 41, 40, 51, 50, 231, 230, 230, 61, 60, 241, 240, 240, 71, 70, 251, 250, 250, 81, 80, 261, 260, 260, 91, 90, 271, 270, 270, 101, 100, 231, 230, 111, 110, 241, 240, 121, 120, 251, 250, 131, 130, 261, 260, 271, 270, 141, 140, 281, 280, 280, 151, 150, 291, 290, 290, 161, 160, 301, 300, 300, 171, 170, 311, 310, 310, 181, 180, 321, 320, 320, 191, 190, 281, 280, 201, 200, 291, 290, 211, 210, 301, 300, 221, 220, 311, 310, 321, 320, 331, 330, 330, 341, 340, 340, 351, 350, 350, 361, 360, 360, 371, 370, 370, 381, 380, 380, 231, 230, 331, 330, 241, 240, 341, 340, 251, 250, 351, 350, 261, 260, 271, 270, 361, 360, 371, 370, 381, 380, 391, 390, 390, 401, 400, 400, 411, 410, 410, 421, 420, 420, 281, 280, 431, 430, 430, 291, 290, 441, 440, 440, 301, 300, 391, 390, 311, 310, 321, 320, 401, 400, 11, 10, 411, 410, 21, 20, 421, 420, 31, 30, 431, 430, 41, 40, 51, 50, 441, 440, 61, 60, 331, 330, 71, 70, 341, 340, 81, 80, 351, 350, 91, 90, 361, 360, 371, 370, 101, 100, 381, 380, 111, 110, 451, 450, 450, 121, 120, 461, 460, 460, 131, 130, 471, 470, 470, 141, 140, 481, 480, 480, 151, 150, 391, 390, 161, 160, 451, 450, 171, 170, 401, 400, 181, 180, 411, 410, 191, 190, 421, 420, 201, 200, 431, 430, 211, 210, 441, 440, 221, 220, 461, 460, 471, 470, 481, 480, 491, 490, 490, 501, 500, 500, 511, 510, 510, 521, 520, 520, 531, 530, 530, 231, 230, 541, 540, 540, 241, 240, 491, 490, 251, 250, 501, 500, 261, 260, 271, 270, 511, 510, 451, 450, 521, 520, 531, 530, 541, 540, 551, 550, 550, 561, 560, 560, 281, 280, 571, 570, 570, 291, 290, 581, 580, 580, 301, 300, 591, 590, 590, 311, 310, 321, 320, 461, 460, 471, 470, 481, 480, 551, 550, 561, 560, 571, 570, 581, 580, 591, 590, 601, 600, 600, 331, 330, 611, 610, 610, 341, 340, 491, 490, 351, 350, 501, 500, 361, 360, 371, 370, 381, 380, 511, 510, 521, 520, 531, 530, 541, 540, 601, 600, 611, 610, 621, 620, 620, 391, 390, 631, 630, 630, 641, 640, 640, 401, 400, 651, 650, 650, 411, 410, 661, 660, 660, 421, 420, 621, 620, 431, 430, 551, 550, 441, 440, 561, 560, 571, 570, 581, 580, 591, 590, 631, 630, 641, 640, 651, 650, 661, 660, 671, 670, 670, 681, 680, 680, 691, 690, 690, 701, 700, 700, 711, 710, 710, 601, 600, 611, 610, 671, 670, 11, 10, 451, 450, 21, 20, 681, 680, 31, 30, 691, 690, 41, 40, 51, 50, 621, 620, 61, 60, 701, 700, 71, 70, 711, 710, 81, 80, 721, 720, 720, 91, 90, 731, 730, 730, 461, 460, 101, 100, 471, 470, 111, 110, 481, 480, 121, 120, 631, 630, 131, 130, 641, 640, 651, 650, 141, 140, 661, 660, 151, 150, 491, 490, 161, 160, 501, 500, 171, 170, 671, 670, 181, 180, 511, 510, 191, 190, 521, 520, 201, 200, 531, 530, 211, 210, 541, 540, 221, 220, 681, 680, 691, 690, 701, 700, 711, 710, 721, 720, 731, 730, 741, 740, 740, 551, 550, 751, 750, 750, 561, 560, 231, 230, 571, 570, 241, 240, 581, 580, 251, 250, 591, 590, 261, 260, 271, 270, 741, 740, 751, 750, 761, 760, 760, 771, 770, 770, 781, 780, 780, 601, 600, 611, 610, 281, 280, 791, 790, 790, 291, 290, 761, 760, 301, 300, 771, 770, 311, 310, 321, 320, 781, 780, 721, 720, 621, 620, 731, 730, 791, 790, 801, 800, 800, 811, 810, 810, 821, 820, 820, 331, 330, 831, 830, 830, 341, 340, 841, 840, 840, 351, 350, 801, 800, 361, 360, 371, 370, 381, 380, 631, 630, 641, 640, 651, 650, 661, 660, 741, 740, 751, 750, 811, 810, 821, 820, 391, 390, 671, 670, 761, 760, 771, 770, 401, 400, 781, 780, 411, 410, 831, 830, 421, 420, 841, 840, 431, 430, 791, 790, 441, 440, 681, 680, 691, 690, 701, 700, 711, 710, 851, 850, 850, 861, 860, 860, 801, 800, 871, 870, 870, 881, 880, 880, 891, 890, 890, 901, 900, 900, 851, 850, 861, 860, 911, 910, 910, 871, 870, 811, 810, 821, 820, 881, 880, 451, 450, 891, 890, 901, 900, 911, 910, 921, 920, 920, 831, 830, 931, 930, 930, 841, 840, 941, 940, 940, 951, 950, 950, 961, 960, 960, 921, 920, 461, 460, 721, 720, 471, 470, 731, 730, 481, 480, 931, 930, 941, 940, 951, 950, 961, 960, 971, 970, 970, 851, 850, 491, 490, 861, 860, 501, 500, 871, 870, 981, 980, 980, 881, 880, 511, 510, 891, 890, 521, 520, 901, 900, 531, 530, 911, 910, 541, 540, 741, 740, 751, 750, 971, 970, 981, 980, 991, 990, 990, 921, 920, 761, 760, 771, 770, 781, 780, 551, 550, 1001, 1000, 1000, 561, 560, 931, 930, 571, 570, 791, 790, 581, 580, 941, 940, 591, 590, 951, 950, 961, 960, 991, 990, 1001, 1000, 801, 800, 1011, 1010, 1010, 1021, 1020, 1020, 1031, 1030, 1030, 601, 600, 611, 610, 1041, 1040, 1040, 1051, 1050, 1050, 1011, 1010, 811, 810, 821, 820, 971, 970, 981, 980, 11, 10, 621, 620, 21, 20, 1021, 1020, 31, 30, 831, 830, 41, 40, 51, 50, 841, 840, 61, 60, 1031, 1030, 71, 70, 1041, 1040, 81, 80, 1051, 1050, 91, 90, 991, 990, 1001, 1000, 101, 100, 631, 630, 111, 110, 641, 640, 121, 120, 651, 650, 131, 130, 661, 660, 851, 850, 141, 140, 671, 670, 151, 150, 861, 860, 161, 160, 871, 870, 171, 170, 881, 880, 181, 180, 891, 890, 191, 190, 901, 900, 201, 200, 681, 680, 211, 210, 691, 690, 221, 220, 701, 700, 711, 710, 911, 910, 921, 920, 931, 930, 1011, 1010, 941, 940, 1021, 1020, 951, 950, 961, 960, 231, 230, 1031, 1030, 241, 240, 1041, 1040, 251, 250, 1051, 1050, 261, 260, 271, 270, 1061, 1060, 1060, 1071, 1070, 1070, 1081, 1080, 1080, 1091, 1090, 1090, 1101, 1100, 1100, 971, 970, 1111, 1110, 1110, 281, 280, 981, 980, 291, 290, 1061, 1060, 301, 300, 721, 720, 311, 310, 321, 320, 731, 730, 1071, 1070, 1081, 1080, 1091, 1090, 1101, 1100, 1111, 1110, 1121, 1120, 1120, 991, 990, 331, 330, 1001, 1000, 341, 340, 1131, 1130, 1130, 351, 350, 1141, 1140, 1140, 361, 360, 371, 370, 381, 380, 741, 740, 751, 750, 1121, 1120, 1131, 1130, 1141, 1140, 761, 760, 771, 770, 781, 780, 391, 390, 1061, 1060, 1151, 1150, 1150, 401, 400, 791, 790, 411, 410, 1071, 1070, 421, 420, 1081, 1080, 431, 430, 1091, 1090, 441, 440, 801, 800, 1101, 1100, 1111, 1110, 1151, 1150, 1011, 1010, 1161, 1160, 1160, 1021, 1020, 1171, 1170, 1170, 1181, 1180, 1180, 1031, 1030, 811, 810, 821, 820, 1041, 1040, 1051, 1050, 1121, 1120, 1131, 1130, 1141, 1140, 1161, 1160, 1171, 1170, 831, 830, 451, 450, 1181, 1180, 841, 840, 1191, 1190, 1190, 1201, 1200, 1200, 1211, 1210, 1210, 1221, 1220, 1220, 1231, 1230, 1230, 1241, 1240, 1240, 1191, 1190, 1251, 1250, 1250, 461, 460, 1151, 1150, 471, 470, 1201, 1200, 481, 480, 1211, 1210, 851, 850, 1221, 1220, 1231, 1230, 1241, 1240, 861, 860, 1251, 1250, 871, 870, 491, 490, 881, 880, 501, 500, 891, 890, 901, 900, 1161, 1160, 511, 510, 1171, 1170, 521, 520, 1181, 1180, 531, 530, 1261, 1260, 1260, 541, 540, 1271, 1270, 1270, 911, 910, 921, 920, 931, 930, 1061, 1060, 941, 940, 1191, 1190, 951, 950, 551, 550, 961, 960, 561, 560, 1071, 1070, 571, 570, 1081, 1080, 581, 580, 1091, 1090, 591, 590, 1101, 1100, 1111, 1110, 1201, 1200, 1211, 1210, 1221, 1220, 1231, 1230, 1241, 1240, 971, 970, 1251, 1250, 601, 600, 611, 610, 981, 980, 1121, 1120, 1131, 1130, 1141, 1140, 1261, 1260, 1271, 1270, 1281, 1280, 1280, 1291, 1290, 1290, 621, 620, 1301, 1300, 1300, 1311, 1310, 1310, 991, 990, 1321, 1320, 1320, 1001, 1000, 1281, 1280, 1291, 1290, 1331, 1330, 1330, 1301, 1300, 1341, 1340, 1340, 1151, 1150, 1311, 1310, 1321, 1320, 631, 630, 1351, 1350, 1350, 641, 640, 1331, 1330, 651, 650, 1341, 1340, 661, 660, 1361, 1360, 1360, 671, 670, 1351, 1350, 1371, 1370, 1370, 1381, 1380, 1380, 1161, 1160, 1261, 1260, 1171, 1170, 1271, 1270, 1181, 1180, 1361, 1360, 681, 680, 1371, 1370, 691, 690, 1381, 1380, 701, 700, 711, 710, 1011, 1010, 1281, 1280, 1021, 1020, 1191, 1190, 1291, 1290, 1031, 1030, 1301, 1300, 1311, 1310, 1041, 1040, 1051, 1050, 1321, 1320, 1331, 1330, 1341, 1340, 1391, 1390, 1390, 1401, 1400, 1400, 1201, 1200, 1211, 1210, 1221, 1220, 1231, 1230, 1241, 1240, 1351, 1350, 1251, 1250, 1391, 1390, 1401, 1400, 1361, 1360, 1411, 1410, 1410, 1371, 1370, 1421, 1420, 1420, 1381, 1380, 721, 720, 1431, 1430, 1430, 1441, 1440, 1440, 731, 730, 1411, 1410, 1451, 1450, 1450, 1421, 1420, 1461, 1460, 1460, 1471, 1470, 1470, 1431, 1430, 1441, 1440, 1481, 1480, 1480, 1451, 1450, 1491, 1490, 1490, 1461, 1460, 1471, 1470, 1501, 1500, 1500, 11, 10, 741, 740, 21, 20, 751, 750, 31, 30, 1391, 1390, 41, 40, 51, 50, 761, 760, 61, 60, 771, 770, 71, 70, 781, 780, 81, 80, 791, 790, 91, 90, 1061, 1060, 1071, 1070, 101, 100, 1081, 1080, 111, 110, 801, 800, 121, 120, 1091, 1090, 131, 130, 1101, 1100, 1111, 1110, 141, 140, 1261, 1260, 151, 150, 1271, 1270, 161, 160, 811, 810, 171, 170, 821, 820, 181, 180, 1121, 1120, 191, 190, 1131, 1130, 201, 200, 1141, 1140, 211, 210, 831, 830, 221, 220, 1281, 1280, 841, 840, 1291, 1290, 1301, 1300, 1311, 1310, 1321, 1320, 1331, 1330, 1151, 1150, 1341, 1340, 1351, 1350, 1361, 1360, 231, 230, 1371, 1370, 241, 240, 1381, 1380, 251, 250, 1401, 1400, 261, 260, 271, 270, 851, 850, 1411, 1410, 1421, 1420, 861, 860, 871, 870, 1161, 1160, 881, 880, 1171, 1170, 891, 890, 281, 280, 901, 900, 291, 290, 1181, 1180, 301, 300, 1391, 1390, 311, 310, 321, 320, 1431, 1430, 911, 910, 921, 920, 931, 930, 1191, 1190, 941, 940, 1441, 1440, 951, 950, 1451, 1450, 331, 330, 961, 960, 341, 340, 1461, 1460, 351, 350, 1201, 1200, 361, 360, 371, 370, 381, 380, 1211, 1210, 1221, 1220, 1231, 1230, 1241, 1240, 1251, 1250, 971, 970, 1471, 1470, 1481, 1480, 391, 390, 1491, 1490, 981, 980, 1501, 1500, 401, 400, 1511, 1510, 1510, 411, 410, 1521, 1520, 1520, 421, 420, 1531, 1530, 1530, 431, 430, 1541, 1540, 1540, 441, 440, 1551, 1550, 1550, 1511, 1510, 1561, 1560, 1560, 991, 990, 1521, 1520, 1001, 1000, 1531, 1530, 1541, 1540, 1571, 1570, 1570, 1551, 1550, 1581, 1580, 1580, 1561, 1560, 1591, 1590, 1590, 1601, 1600, 1600, 1401, 1400, 1611, 1610, 1610, 1571, 1570, 1621, 1620, 1620, 1581, 1580, 1411, 1410, 1421, 1420, 1481, 1480, 451, 450, 1491, 1490, 1501, 1500, 1591, 1590, 1601, 1600, 1611, 1610, 1621, 1620, 1631, 1630, 1630, 1641, 1640, 1640, 1651, 1650, 1650, 1511, 1510, 461, 460, 1431, 1430, 471, 470, 1521, 1520, 481, 480, 1011, 1010, 1021, 1020, 1261, 1260, 1271, 1270, 1031, 1030, 1441, 1440, 491, 490, 1041, 1040, 501, 500, 1051, 1050, 1451, 1450, 1461, 1460, 511, 510, 1531, 1530, 521, 520, 1541, 1540, 531, 530, 1281, 1280, 541, 540, 1291, 1290, 1301, 1300, 1311, 1310, 1321, 1320, 1331, 1330, 1471, 1470, 1341, 1340, 1351, 1350, 1361, 1360, 551, 550, 1371, 1370, 561, 560, 1381, 1380, 571, 570, 1551, 1550, 581, 580, 1561, 1560, 591, 590, 1571, 1570, 1581, 1580, 1591, 1590, 1601, 1600, 1611, 1610, 1621, 1620, 1631, 1630, 1641, 1640, 1651, 1650, 601, 600, 611, 610, 1661, 1660, 1660, 1391, 1390, 1671, 1670, 1670, 1481, 1480, 1681, 1680, 1680, 1491, 1490, 1501, 1500, 1691, 1690, 1690, 621, 620, 1661, 1660, 1701, 1700, 1700, 1671, 1670, 1711, 1710, 1710, 1681, 1680, 1721, 1720, 1720, 1691, 1690, 1511, 1510, 1061, 1060, 1071, 1070, 1081, 1080, 1521, 1520, 1701, 1700, 631, 630, 1091, 1090, 641, 640, 1101, 1100, 651, 650, 1111, 1110, 661, 660, 1631, 1630, 671, 670, 1641, 1640, 1651, 1650, 1711, 1710, 1121, 1120, 1531, 1530, 1131, 1130, 1541, 1540, 1141, 1140, 1721, 1720, 1661, 1660, 1731, 1730, 1730, 681, 680, 1671, 1670, 691, 690, 1681, 1680, 701, 700, 711, 710, 1691, 1690, 1731, 1730, 1151, 1150, 1741, 1740, 1740, 1701, 1700, 1751, 1750, 1750, 1401, 1400, 1551, 1550, 1561, 1560, 1761, 1760, 1760, 1411, 1410, 1421, 1420, 1571, 1570, 1581, 1580, 1591, 1590, 1601, 1600, 1161, 1160, 1611, 1610, 1171, 1170, 1621, 1620, 1711, 1710, 1721, 1720, 1741, 1740, 1181, 1180, 1751, 1750, 1761, 1760, 1771, 1770, 1770, 1431, 1430, 721, 720, 1781, 1780, 1780, 1791, 1790, 1790, 731, 730, 1731, 1730, 1191, 1190, 1771, 1770, 1801, 1800, 1800, 1441, 1440, 1811, 1810, 1810, 1781, 1780, 1791, 1790, 1821, 1820, 1820, 1451, 1450, 1461, 1460, 1201, 1200, 1801, 1800, 1811, 1810, 1831, 1830, 1830, 1211, 1210, 1221, 1220, 741, 740, 1231, 1230, 751, 750, 1241, 1240, 1251, 1250, 1471, 1470, 761, 760, 1631, 1630, 771, 770, 1641, 1640, 781, 780, 1651, 1650, 791, 790, 1661, 1660, 1671, 1670, 1741, 1740, 1681, 1680, 1751, 1750, 801, 800, 1761, 1760, 1691, 1690, 1771, 1770, 1781, 1780, 1791, 1790, 1821, 1820, 1701, 1700, 1801, 1800, 1811, 1810, 811, 810, 1831, 1830, 821, 820, 1841, 1840, 1840, 1851, 1850, 1850, 1481, 1480, 1861, 1860, 1860, 831, 830, 1491, 1490, 1501, 1500, 841, 840, 1841, 1840, 1851, 1850, 1711, 1710, 1721, 1720, 1861, 1860, 1871, 1870, 1870, 1511, 1510, 1881, 1880, 1880, 1891, 1890, 1890, 1521, 1520, 1901, 1900, 1900, 1731, 1730, 1871, 1870, 1821, 1820, 1911, 1910, 1910, 851, 850, 1881, 1880, 1891, 1890, 861, 860, 871, 870, 1261, 1260, 881, 880, 1271, 1270, 891, 890, 1831, 1830, 901, 900, 1531, 1530, 1541, 1540, 1901, 1900, 1911, 1910, 1841, 1840, 1851, 1850, 1281, 1280, 911, 910, 921, 920, 931, 930, 1291, 1290, 941, 940, 1301, 1300, 951, 950, 1311, 1310, 1321, 1320, 961, 960, 1331, 1330, 1341, 1340, 1351, 1350, 1361, 1360, 1371, 1370, 1381, 1380, 1551, 1550, 1561, 1560, 1571, 1570, 1581, 1580, 1591, 1590, 1601, 1600, 971, 970, 1611, 1610, 1621, 1620, 1741, 1740, 981, 980, 1751, 1750, 1761, 1760, 1391, 1390, 1771, 1770, 11, 10, 1781, 1780, 21, 20, 1791, 1790, 31, 30, 1801, 1800, 41, 40, 51, 50, 991, 990, 61, 60, 1001, 1000, 71, 70, 1811, 1810, 81, 80, 1861, 1860, 91, 90, 1871, 1870, 1881, 1880, 101, 100, 1821, 1820, 111, 110, 1891, 1890, 121, 120, 1901, 1900, 131, 130, 1911, 1910, 1921, 1920, 1920, 141, 140, 1931, 1930, 1930, 151, 150, 1831, 1830, 161, 160, 1941, 1940, 1940, 171, 170, 1631, 1630, 181, 180, 1641, 1640, 191, 190, 1651, 1650, 201, 200, 1661, 1660, 211, 210, 1671, 1670, 221, 220, 1681, 1680, 1691, 1690, 1841, 1840, 1851, 1850, 1011, 1010, 1021, 1020, 1701, 1700, 1921, 1920, 1031, 1030, 1931, 1930, 231, 230, 1041, 1040, 241, 240, 1051, 1050, 251, 250, 1401, 1400, 261, 260, 271, 270, 1411, 1410, 1421, 1420, 1941, 1940, 1711, 1710, 1721, 1720, 1951, 1950, 1950, 1961, 1960, 1960, 281, 280, 1971, 1970, 1970, 291, 290, 1981, 1980, 1980, 301, 300, 1991, 1990, 1990, 311, 310, 321, 320, 1431, 1430, 1731, 1730, 1951, 1950, 1961, 1960, 1921, 1920, 1971, 1970, 1931, 1930, 1981, 1980, 1441, 1440, 331, 330, 1861, 1860, 341, 340, 1871, 1870, 351, 350, 1451, 1450, 361, 360, 371, 370, 381, 380, 1461, 1460, 1881, 1880, 1891, 1890, 1901, 1900, 1911, 1910, 1941, 1940, 1991, 1990, 2001, 2000, 2000, 391, 390, 1471, 1470, 401, 400, 411, 410, 421, 420, 1061, 1060, 431, 430, 1071, 1070, 441, 440, 1081, 1080, 1091, 1090, 1951, 1950, 1101, 1100, 1961, 1960, 1111, 1110, 1971, 1970, 1981, 1980, 1741, 1740, 2001, 2000, 1751, 1750, 1761, 1760, 1121, 1120, 1131, 1130, 1771, 1770, 1141, 1140, 1481, 1480, 1781, 1780, 1791, 1790, 1491, 1490, 1501, 1500, 451, 450, 1801, 1800, 1991, 1990, 1811, 1810, 1151, 1150, 1511, 1510, 1821, 1820, 1521, 1520, 461, 460, 471, 470, 1921, 1920, 481, 480, 1931, 1930, 1831, 1830, 2001, 2000, 1161, 1160, 1171, 1170, 491, 490, 501, 500, 1181, 1180, 1531, 1530, 511, 510, 1541, 1540, 521, 520, 531, 530, 1841, 1840, 541, 540, 1851, 1850, 1941, 1940, 1191, 1190, 551, 550, 561, 560, 1201, 1200, 571, 570, 1951, 1950, 581, 580, 1551, 1550, 1211, 1210, 591, 590, 1221, 1220, 1231, 1230, 1241, 1240, 1251, 1250, 1561, 1560, 1571, 1570, 1581, 1580, 1591, 1590, 1601, 1600, 601, 600, 611, 610, 1611, 1610, 1621, 1620, 1961, 1960, 1971, 1970, 1981, 1980, 1991, 1990, 621, 620, 1861, 1860, 1871, 1870, 1881, 1880, 1891, 1890, 1901, 1900, 1911, 1910, 631, 630, 2001, 2000, 641, 640, 651, 650, 661, 660, 1631, 1630, 671, 670, 1641, 1640, 1651, 1650, 1661, 1660, 1671, 1670, 1681, 1680, 1691, 1690, 681, 680, 691, 690, 701, 700, 711, 710, 1261, 1260, 1701, 1700, 1271, 1270, 1281, 1280, 1711, 1710, 1721, 1720, 1291, 1290, 1301, 1300, 1311, 1310, 1321, 1320, 1331, 1330, 1341, 1340, 1351, 1350, 1361, 1360, 1371, 1370, 1381, 1380, 721, 720, 1731, 1730, 1921, 1920, 731, 730, 1931, 1930, 1391, 1390, 741, 740, 1941, 1940, 751, 750, 761, 760, 771, 770, 781, 780, 791, 790, 1951, 1950, 801, 800, 1741, 1740, 1751, 1750, 811, 810, 1761, 1760, 821, 820, 1771, 1770, 1781, 1780, 1791, 1790, 1961, 1960, 831, 830, 1971, 1970, 1981, 1980, 841, 840, 1801, 1800, 1811, 1810, 1991, 1990, 1821, 1820, 1401, 1400, 1411, 1410, 1421, 1420, 851, 850, 1831, 1830, 2001, 2000, 861, 860, 871, 870, 881, 880, 891, 890, 901, 900, 1431, 1430, 1841, 1840, 911, 910, 921, 920, 931, 930, 1851, 1850, 941, 940, 1441, 1440, 951, 950, 961, 960, 1451, 1450, 1461, 1460, 971, 970, 1471, 1470, 981, 980, 991, 990, 1001, 1000, 1861, 1860, 1871, 1870, 1881, 1880, 1481, 1480, 1891, 1890, 1901, 1900, 1491, 1490, 1501, 1500, 1911, 1910, 1511, 1510, 1521, 1520, 1011, 1010, 1021, 1020, 1031, 1030, 1041, 1040, 1051, 1050, 1531, 1530, 1541, 1540, 1551, 1550, 1921, 1920, 1931, 1930, 1561, 1560, 1571, 1570, 1581, 1580, 1591, 1590, 1601, 1600, 1611, 1610, 1621, 1620, 1941, 1940, 1061, 1060, 1071, 1070, 1081, 1080, 1091, 1090, 1951, 1950, 1101, 1100, 1111, 1110, 1631, 1630, 1121, 1120, 11, 10, 1131, 1130, 21, 20, 1141, 1140, 31, 30, 1641, 1640, 41, 40, 51, 50, 1651, 1650, 61, 60, 1661, 1660, 71, 70, 1671, 1670, 81, 80, 1681, 1680, 91, 90, 1151, 1150, 1691, 1690, 101, 100, 1701, 1700, 111, 110, 1961, 1960, 121, 120, 1971, 1970, 131, 130, 1981, 1980, 1991, 1990, 141, 140, 1711, 1710, 151, 150, 1721, 1720, 161, 160, 2001, 2000, 171, 170, 1161, 1160, 181, 180, 1171, 1170, 191, 190, 201, 200, 211, 210, 1181, 1180, 221, 220, 1731, 1730, 1191, 1190, 231, 230, 241, 240, 251, 250, 1201, 1200, 261, 260, 271, 270, 1211, 1210, 1221, 1220, 1231, 1230, 1241, 1240, 1251, 1250, 281, 280, 291, 290, 301, 300, 311, 310, 321, 320, 1741, 1740, 331, 330, 1751, 1750, 341, 340, 1761, 1760, 351, 350, 1771, 1770, 361, 360, 371, 370, 381, 380, 1781, 1780, 1791, 1790, 1801, 1800, 1811, 1810, 391, 390, 1821, 1820, 401, 400, 411, 410, 421, 420, 431, 430, 441, 440, 1831, 1830, 1261, 1260, 1271, 1270, 1841, 1840, 1281, 1280, 451, 450, 1851, 1850, 1291, 1290, 1301, 1300, 1311, 1310, 1321, 1320, 1331, 1330, 1341, 1340, 1351, 1350, 1361, 1360, 461, 460, 1371, 1370, 471, 470, 1381, 1380, 481, 480, 491, 490, 501, 500, 1391, 1390, 511, 510, 521, 520, 531, 530, 541, 540, 1861, 1860, 1871, 1870, 551, 550, 1881, 1880, 561, 560, 1891, 1890, 571, 570, 1901, 1900, 581, 580, 1911, 1910, 591, 590, 601, 600, 611, 610, 621, 620, 1401, 1400, 631, 630, 1411, 1410, 641, 640, 1421, 1420, 651, 650, 661, 660, 671, 670, 1431, 1430, 1921, 1920, 681, 680, 1931, 1930, 691, 690, 701, 700, 711, 710, 1441, 1440, 1451, 1450, 1461, 1460, 1941, 1940, 1471, 1470, 721, 720, 731, 730, 1481, 1480, 1491, 1490, 741, 740, 1501, 1500, 751, 750, 761, 760, 771, 770, 1511, 1510, 781, 780, 791, 790, 1521, 1520, 801, 800, 811, 810, 821, 820, 831, 830, 841, 840, 851, 850, 861, 860, 871, 870, 881, 880, 891, 890, 901, 900, 911, 910, 921, 920, 931, 930, 941, 940, 951, 950, 961, 960, 971, 970, 981, 980}
+	// Advanced multi = 40
+	Advanced = []uint16{11, 10, 10, 21, 20, 20, 31, 30, 30, 41, 40, 40, 51, 50, 50, 61, 60, 60, 71, 70, 70, 81, 80, 80, 91, 90, 90, 101, 100, 100, 111, 110, 110, 121, 120, 120, 131, 130, 130, 141, 140, 140, 11, 10, 151, 150, 150, 21, 20, 161, 160, 160, 31, 30, 171, 170, 170, 41, 40, 181, 180, 180, 51, 50, 191, 190, 190, 61, 60, 201, 200, 200, 71, 70, 211, 210, 210, 81, 80, 221, 220, 220, 91, 90, 231, 230, 230, 101, 100, 241, 240, 240, 111, 110, 251, 250, 250, 121, 120, 261, 260, 260, 131, 130, 271, 270, 270, 141, 140, 281, 280, 280, 151, 150, 291, 290, 290, 161, 160, 301, 300, 300, 171, 170, 311, 310, 310, 181, 180, 321, 320, 320, 191, 190, 331, 330, 330, 201, 200, 341, 340, 340, 211, 210, 351, 350, 350, 221, 220, 361, 360, 360, 231, 230, 371, 370, 370, 241, 240, 381, 380, 380, 251, 250, 391, 390, 390, 261, 260, 401, 400, 400, 271, 270, 411, 410, 410, 281, 280, 421, 420, 420, 291, 290, 431, 430, 430, 301, 300, 441, 440, 440, 311, 310, 451, 450, 450, 321, 320, 461, 460, 460, 331, 330, 471, 470, 470, 341, 340, 481, 480, 480, 351, 350, 491, 490, 490, 361, 360, 501, 500, 500, 371, 370, 511, 510, 510, 381, 380, 521, 520, 520, 391, 390, 531, 530, 530, 401, 400, 541, 540, 540, 411, 410, 551, 550, 550, 421, 420, 561, 560, 560, 431, 430, 571, 570, 570, 441, 440, 581, 580, 580, 451, 450, 591, 590, 590, 461, 460, 601, 600, 600, 471, 470, 611, 610, 610, 481, 480, 621, 620, 620, 491, 490, 631, 630, 630, 501, 500, 641, 640, 640, 511, 510, 651, 650, 650, 521, 520, 661, 660, 660, 531, 530, 541, 540, 11, 10, 551, 550, 21, 20, 561, 560, 31, 30, 571, 570, 41, 40, 581, 580, 51, 50, 591, 590, 61, 60, 601, 600, 71, 70, 611, 610, 81, 80, 621, 620, 91, 90, 631, 630, 101, 100, 641, 640, 111, 110, 651, 650, 121, 120, 661, 660, 131, 130, 671, 670, 670, 141, 140, 681, 680, 680, 151, 150, 691, 690, 690, 161, 160, 701, 700, 700, 171, 170, 711, 710, 710, 181, 180, 721, 720, 720, 191, 190, 731, 730, 730, 201, 200, 741, 740, 740, 211, 210, 751, 750, 750, 221, 220, 761, 760, 760, 231, 230, 771, 770, 770, 241, 240, 781, 780, 780, 251, 250, 791, 790, 790, 261, 260, 801, 800, 800, 271, 270, 671, 670, 281, 280, 681, 680, 291, 290, 691, 690, 301, 300, 701, 700, 311, 310, 711, 710, 321, 320, 721, 720, 331, 330, 731, 730, 341, 340, 741, 740, 351, 350, 751, 750, 361, 360, 761, 760, 371, 370, 771, 770, 381, 380, 781, 780, 391, 390, 791, 790, 401, 400, 801, 800, 411, 410, 811, 810, 810, 421, 420, 821, 820, 820, 431, 430, 831, 830, 830, 441, 440, 841, 840, 840, 451, 450, 851, 850, 850, 461, 460, 861, 860, 860, 471, 470, 871, 870, 870, 481, 480, 881, 880, 880, 491, 490, 891, 890, 890, 501, 500, 901, 900, 900, 511, 510, 911, 910, 910, 521, 520, 531, 530, 921, 920, 920, 541, 540, 931, 930, 930, 551, 550, 811, 810, 561, 560, 821, 820, 571, 570, 831, 830, 581, 580, 841, 840, 591, 590, 851, 850, 601, 600, 861, 860, 611, 610, 871, 870, 621, 620, 881, 880, 631, 630, 891, 890, 641, 640, 901, 900, 651, 650, 911, 910, 661, 660, 921, 920, 931, 930, 941, 940, 940, 951, 950, 950, 961, 960, 960, 971, 970, 970, 981, 980, 980, 991, 990, 990, 1001, 1000, 1000, 1011, 1010, 1010, 1021, 1020, 1020, 1031, 1030, 1030, 1041, 1040, 1040, 1051, 1050, 1050, 1061, 1060, 1060, 671, 670, 1071, 1070, 1070, 681, 680, 941, 940, 691, 690, 951, 950, 701, 700, 961, 960, 711, 710, 971, 970, 721, 720, 981, 980, 731, 730, 991, 990, 741, 740, 1001, 1000, 751, 750, 1011, 1010, 761, 760, 1021, 1020, 771, 770, 1031, 1030, 781, 780, 1041, 1040, 791, 790, 1051, 1050, 801, 800, 1061, 1060, 1071, 1070, 1081, 1080, 1080, 1091, 1090, 1090, 1101, 1100, 1100, 1111, 1110, 1110, 1121, 1120, 1120, 1131, 1130, 1130, 1141, 1140, 1140, 1151, 1150, 1150, 1161, 1160, 1160, 1171, 1170, 1170, 1181, 1180, 1180, 1191, 1190, 1190, 1201, 1200, 1200, 811, 810, 1211, 1210, 1210, 821, 820, 1081, 1080, 831, 830, 1091, 1090, 841, 840, 1101, 1100, 851, 850, 1111, 1110, 861, 860, 1121, 1120, 871, 870, 1131, 1130, 881, 880, 1141, 1140, 891, 890, 1151, 1150, 901, 900, 1161, 1160, 911, 910, 1171, 1170, 921, 920, 931, 930, 11, 10, 1181, 1180, 21, 20, 1191, 1190, 31, 30, 1201, 1200, 41, 40, 1211, 1210, 51, 50, 1221, 1220, 1220, 61, 60, 1231, 1230, 1230, 71, 70, 1241, 1240, 1240, 81, 80, 1251, 1250, 1250, 91, 90, 1261, 1260, 1260, 101, 100, 1271, 1270, 1270, 111, 110, 1281, 1280, 1280, 121, 120, 1291, 1290, 1290, 131, 130, 1301, 1300, 1300, 141, 140, 941, 940, 151, 150, 951, 950, 161, 160, 961, 960, 171, 170, 971, 970, 181, 180, 981, 980, 191, 190, 991, 990, 201, 200, 1001, 1000, 211, 210, 1011, 1010, 221, 220, 1021, 1020, 231, 230, 1031, 1030, 241, 240, 1041, 1040, 251, 250, 1051, 1050, 261, 260, 1061, 1060, 271, 270, 1071, 1070, 281, 280, 1221, 1220, 291, 290, 1231, 1230, 301, 300, 1241, 1240, 311, 310, 1251, 1250, 321, 320, 1261, 1260, 331, 330, 1271, 1270, 341, 340, 1281, 1280, 351, 350, 1291, 1290, 361, 360, 1301, 1300, 371, 370, 1081, 1080, 381, 380, 1091, 1090, 391, 390, 1101, 1100, 401, 400, 1111, 1110, 411, 410, 1121, 1120, 421, 420, 1131, 1130, 431, 430, 1141, 1140, 441, 440, 1151, 1150, 451, 450, 1161, 1160, 461, 460, 1171, 1170, 471, 470, 1181, 1180, 481, 480, 1191, 1190, 491, 490, 1201, 1200, 501, 500, 1211, 1210, 511, 510, 1311, 1310, 1310, 521, 520, 531, 530, 1321, 1320, 1320, 541, 540, 1331, 1330, 1330, 551, 550, 1341, 1340, 1340, 561, 560, 1351, 1350, 1350, 571, 570, 1361, 1360, 1360, 581, 580, 1371, 1370, 1370, 591, 590, 1381, 1380, 1380, 601, 600, 1391, 1390, 1390, 611, 610, 1401, 1400, 1400, 621, 620, 1411, 1410, 1410, 631, 630, 1421, 1420, 1420, 641, 640, 1431, 1430, 1430, 651, 650, 1311, 1310, 661, 660, 1321, 1320, 1441, 1440, 1440, 1331, 1330, 1451, 1450, 1450, 1341, 1340, 1461, 1460, 1460, 1351, 1350, 1471, 1470, 1470, 1361, 1360, 1481, 1480, 1480, 1371, 1370, 1491, 1490, 1490, 1381, 1380, 1501, 1500, 1500, 1391, 1390, 1511, 1510, 1510, 1401, 1400, 1521, 1520, 1520, 1411, 1410, 1531, 1530, 1530, 1421, 1420, 1221, 1220, 1431, 1430, 1231, 1230, 1541, 1540, 1540, 1241, 1240, 1551, 1550, 1550, 1251, 1250, 671, 670, 1261, 1260, 681, 680, 1271, 1270, 691, 690, 1281, 1280, 701, 700, 1291, 1290, 711, 710, 1301, 1300, 721, 720, 1441, 1440, 731, 730, 1451, 1450, 741, 740, 1461, 1460, 751, 750, 1471, 1470, 761, 760, 1481, 1480, 771, 770, 1491, 1490, 781, 780, 1501, 1500, 791, 790, 1511, 1510, 801, 800, 1521, 1520, 1531, 1530, 1541, 1540, 1551, 1550, 1561, 1560, 1560, 1571, 1570, 1570, 1581, 1580, 1580, 1591, 1590, 1590, 1601, 1600, 1600, 1611, 1610, 1610, 1621, 1620, 1620, 1631, 1630, 1630, 1641, 1640, 1640, 1651, 1650, 1650, 1661, 1660, 1660, 1671, 1670, 1670, 811, 810, 1681, 1680, 1680, 821, 820, 1691, 1690, 1690, 831, 830, 1561, 1560, 841, 840, 1571, 1570, 851, 850, 1581, 1580, 861, 860, 1591, 1590, 871, 870, 1601, 1600, 881, 880, 1611, 1610, 891, 890, 1311, 1310, 901, 900, 1321, 1320, 911, 910, 1331, 1330, 921, 920, 1341, 1340, 931, 930, 1351, 1350, 1361, 1360, 1371, 1370, 1381, 1380, 1391, 1390, 1621, 1620, 1401, 1400, 1631, 1630, 1411, 1410, 1641, 1640, 1421, 1420, 1651, 1650, 1431, 1430, 1661, 1660, 1671, 1670, 1681, 1680, 1691, 1690, 1701, 1700, 1700, 1711, 1710, 1710, 1721, 1720, 1720, 1731, 1730, 1730, 941, 940, 1741, 1740, 1740, 951, 950, 1751, 1750, 1750, 961, 960, 1761, 1760, 1760, 971, 970, 1441, 1440, 981, 980, 1451, 1450, 991, 990, 1461, 1460, 1001, 1000, 1471, 1470, 1011, 1010, 1481, 1480, 1021, 1020, 1491, 1490, 1031, 1030, 1501, 1500, 1041, 1040, 1511, 1510, 1051, 1050, 1521, 1520, 1061, 1060, 1531, 1530, 1071, 1070, 1541, 1540, 1551, 1550, 1701, 1700, 1711, 1710, 1721, 1720, 1731, 1730, 1741, 1740, 1751, 1750, 1761, 1760, 1771, 1770, 1770, 1781, 1780, 1780, 1791, 1790, 1790, 1801, 1800, 1800, 1081, 1080, 1811, 1810, 1810, 1091, 1090, 1821, 1820, 1820, 1101, 1100, 1561, 1560, 1111, 1110, 1571, 1570, 1121, 1120, 1581, 1580, 1131, 1130, 1591, 1590, 1141, 1140, 1601, 1600, 1151, 1150, 1611, 1610, 1161, 1160, 1771, 1770, 1171, 1170, 1781, 1780, 1181, 1180, 1791, 1790, 1191, 1190, 1801, 1800, 1201, 1200, 1811, 1810, 1211, 1210, 1821, 1820, 1831, 1830, 1830, 1621, 1620, 1841, 1840, 1840, 1631, 1630, 1851, 1850, 1850, 1641, 1640, 1861, 1860, 1860, 1651, 1650, 1871, 1870, 1870, 1661, 1660, 1671, 1670, 1681, 1680, 1691, 1690, 1881, 1880, 1880, 1891, 1890, 1890, 1901, 1900, 1900, 1911, 1910, 1910, 1921, 1920, 1920, 1931, 1930, 1930, 1831, 1830, 1941, 1940, 1940, 1841, 1840, 1951, 1950, 1950, 1851, 1850, 1961, 1960, 1960, 1861, 1860, 1971, 1970, 1970, 1871, 1870, 1981, 1980, 1980, 1991, 1990, 1990, 2001, 2000, 2000, 1881, 1880, 1891, 1890, 1901, 1900, 1911, 1910, 1221, 1220, 1921, 1920, 1231, 1230, 1931, 1930, 1241, 1240, 1941, 1940, 1251, 1250, 1701, 1700, 1261, 1260, 1711, 1710, 1271, 1270, 1721, 1720, 1281, 1280, 1731, 1730, 1291, 1290, 1741, 1740, 1301, 1300, 1751, 1750, 1761, 1760, 1951, 1950, 1961, 1960, 1971, 1970, 1981, 1980, 1991, 1990, 2001, 2000, 11, 10, 21, 20, 31, 30, 41, 40, 1771, 1770, 51, 50, 1781, 1780, 61, 60, 1791, 1790, 71, 70, 1801, 1800, 81, 80, 1811, 1810, 91, 90, 1821, 1820, 101, 100, 111, 110, 121, 120, 131, 130, 141, 140, 151, 150, 161, 160, 171, 170, 181, 180, 191, 190, 201, 200, 211, 210, 1831, 1830, 221, 220, 1841, 1840, 231, 230, 1311, 1310, 241, 240, 1321, 1320, 251, 250, 1331, 1330, 261, 260, 1341, 1340, 271, 270, 1351, 1350, 281, 280, 1361, 1360, 291, 290, 1371, 1370, 301, 300, 1381, 1380, 311, 310, 1391, 1390, 321, 320, 1401, 1400, 331, 330, 1411, 1410, 341, 340, 1421, 1420, 351, 350, 1431, 1430, 361, 360, 1851, 1850, 371, 370, 1861, 1860, 381, 380, 1871, 1870, 391, 390, 1881, 1880, 401, 400, 1441, 1440, 411, 410, 1451, 1450, 421, 420, 1461, 1460, 431, 430, 1471, 1470, 441, 440, 1481, 1480, 451, 450, 1491, 1490, 461, 460, 1501, 1500, 471, 470, 1511, 1510, 481, 480, 1521, 1520, 491, 490, 1531, 1530, 501, 500, 1541, 1540, 511, 510, 1551, 1550, 521, 520, 531, 530, 1891, 1890, 541, 540, 1901, 1900, 551, 550, 1911, 1910, 561, 560, 1921, 1920, 571, 570, 1931, 1930, 581, 580, 1561, 1560, 591, 590, 1571, 1570, 601, 600, 1581, 1580, 611, 610, 1591, 1590, 621, 620, 1601, 1600, 631, 630, 1611, 1610, 641, 640, 1941, 1940, 651, 650, 1951, 1950, 661, 660, 1961, 1960, 1971, 1970, 1981, 1980, 1621, 1620, 1991, 1990, 1631, 1630, 2001, 2000, 1641, 1640, 1651, 1650, 1661, 1660, 1671, 1670, 1681, 1680, 1691, 1690, 671, 670, 681, 680, 691, 690, 701, 700, 711, 710, 721, 720, 731, 730, 741, 740, 751, 750, 761, 760, 771, 770, 781, 780, 791, 790, 801, 800, 1701, 1700, 1711, 1710, 1721, 1720, 1731, 1730, 1741, 1740, 1751, 1750, 1761, 1760, 811, 810, 821, 820, 831, 830, 841, 840, 851, 850, 861, 860, 1771, 1770, 871, 870, 1781, 1780, 881, 880, 1791, 1790, 891, 890, 1801, 1800, 901, 900, 1811, 1810, 911, 910, 1821, 1820, 921, 920, 931, 930, 1831, 1830, 1841, 1840, 941, 940, 951, 950, 961, 960, 971, 970, 981, 980, 991, 990, 1001, 1000, 1011, 1010, 1021, 1020, 1031, 1030, 1041, 1040, 1051, 1050, 1061, 1060, 1071, 1070, 1851, 1850, 1861, 1860, 1871, 1870, 1881, 1880, 1081, 1080, 1091, 1090, 1101, 1100, 1111, 1110, 1121, 1120, 1131, 1130, 1141, 1140, 1151, 1150, 1161, 1160, 1171, 1170, 1181, 1180, 1191, 1190, 1891, 1890, 1201, 1200, 1901, 1900, 1211, 1210, 1911, 1910, 1921, 1920, 1931, 1930, 1941, 1940, 1951, 1950, 1961, 1960, 1971, 1970, 1981, 1980, 1991, 1990, 2001, 2000, 1221, 1220, 1231, 1230, 1241, 1240, 1251, 1250, 1261, 1260, 1271, 1270, 1281, 1280, 1291, 1290, 1301, 1300, 1311, 1310, 1321, 1320, 1331, 1330, 1341, 1340, 1351, 1350, 1361, 1360, 1371, 1370, 1381, 1380, 1391, 1390, 1401, 1400, 1411, 1410, 1421, 1420, 1431, 1430, 1441, 1440, 1451, 1450, 1461, 1460, 1471, 1470, 1481, 1480, 1491, 1490, 1501, 1500, 1511, 1510, 1521, 1520, 1531, 1530, 1541, 1540, 1551, 1550, 1561, 1560, 1571, 1570, 1581, 1580, 1591, 1590, 1601, 1600, 1611, 1610, 1621, 1620, 1631, 1630, 1641, 1640, 1651, 1650, 1661, 1660, 1671, 1670, 1681, 1680, 1691, 1690, 1701, 1700, 1711, 1710, 1721, 1720, 1731, 1730, 1741, 1740, 1751, 1750, 1761, 1760, 1771, 1770, 1781, 1780, 1791, 1790, 1801, 1800, 1811, 1810, 1821, 1820, 11, 10, 21, 20, 31, 30, 41, 40, 51, 50, 61, 60, 71, 70, 81, 80, 91, 90, 1831, 1830, 101, 100, 1841, 1840, 111, 110, 121, 120, 131, 130, 141, 140, 151, 150, 161, 160, 171, 170, 181, 180, 191, 190, 201, 200, 211, 210, 221, 220, 231, 230, 241, 240, 251, 250, 261, 260, 271, 270, 1851, 1850, 281, 280, 1861, 1860, 291, 290, 1871, 1870, 301, 300, 1881, 1880, 311, 310, 321, 320, 331, 330, 341, 340, 351, 350, 361, 360, 371, 370, 381, 380, 391, 390, 401, 400, 411, 410, 421, 420, 431, 430, 441, 440, 451, 450, 461, 460, 1891, 1890, 471, 470, 1901, 1900, 481, 480, 1911, 1910, 491, 490, 1921, 1920, 501, 500, 1931, 1930, 511, 510, 521, 520, 531, 530, 541, 540, 551, 550, 561, 560, 571, 570, 581, 580, 1941, 1940, 591, 590, 1951, 1950, 601, 600, 1961, 1960, 611, 610, 1971, 1970, 621, 620, 1981, 1980, 631, 630, 1991, 1990, 641, 640, 2001, 2000, 651, 650, 661, 660, 671, 670, 681, 680, 691, 690, 701, 700, 711, 710, 721, 720, 731, 730, 741, 740, 751, 750, 761, 760, 771, 770, 781, 780, 791, 790, 801, 800, 811, 810, 821, 820, 831, 830, 841, 840, 851, 850, 861, 860, 871, 870, 881, 880, 891, 890, 901, 900, 911, 910, 921, 920, 931, 930, 941, 940, 951, 950, 961, 960, 971, 970, 981, 980, 991, 990, 1001, 1000, 1011, 1010, 1021, 1020, 1031, 1030, 1041, 1040, 1051, 1050, 1061, 1060, 1071, 1070, 1081, 1080, 1091, 1090, 1101, 1100, 1111, 1110, 1121, 1120, 1131, 1130, 1141, 1140, 1151, 1150, 1161, 1160, 1171, 1170, 1181, 1180, 1191, 1190, 1201, 1200, 1211, 1210, 1221, 1220, 1231, 1230, 1241, 1240, 1251, 1250, 1261, 1260, 1271, 1270, 1281, 1280, 1291, 1290, 1301, 1300, 1311, 1310, 1321, 1320, 1331, 1330, 1341, 1340, 1351, 1350, 1361, 1360, 1371, 1370, 1381, 1380, 1391, 1390, 1401, 1400, 1411, 1410, 1421, 1420, 1431, 1430, 1441, 1440, 1451, 1450, 1461, 1460, 1471, 1470, 1481, 1480, 1491, 1490, 1501, 1500, 1511, 1510, 1521, 1520, 1531, 1530, 1541, 1540, 1551, 1550, 1561, 1560, 1571, 1570, 1581, 1580, 1591, 1590, 1601, 1600, 1611, 1610, 1621, 1620, 1631, 1630, 1641, 1640, 1651, 1650, 1661, 1660, 1671, 1670, 1681, 1680, 1691, 1690, 1701, 1700, 1711, 1710, 1721, 1720, 1731, 1730, 1741, 1740, 1751, 1750, 1761, 1760, 1771, 1770, 1781, 1780, 1791, 1790, 1801, 1800, 1811, 1810, 1821, 1820, 1831, 1830, 1841, 1840, 1851, 1850, 1861, 1860, 1871, 1870, 1881, 1880, 1891, 1890, 1901, 1900, 1911, 1910, 1921, 1920, 1931, 1930, 1941, 1940, 1951, 1950, 1961, 1960, 1971, 1970, 1981, 1980, 1991, 1990, 2001, 2000, 11, 10, 21, 20, 31, 30, 41, 40, 51, 50, 61, 60, 71, 70, 81, 80, 91, 90, 101, 100, 111, 110, 121, 120, 131, 130, 141, 140, 151, 150, 161, 160, 171, 170, 181, 180, 191, 190, 201, 200, 211, 210, 221, 220, 231, 230, 241, 240, 251, 250, 261, 260, 271, 270, 281, 280, 291, 290, 301, 300, 311, 310, 321, 320, 331, 330, 341, 340, 351, 350, 361, 360, 371, 370, 381, 380, 391, 390, 401, 400, 411, 410, 421, 420, 431, 430, 441, 440, 451, 450, 461, 460, 471, 470, 481, 480, 491, 490, 501, 500, 511, 510, 521, 520, 531, 530, 541, 540, 551, 550, 561, 560, 571, 570, 581, 580, 591, 590, 601, 600, 611, 610, 621, 620, 631, 630, 641, 640, 651, 650, 661, 660}
+	// Review each phrase is only repeated one time
+	Review = []uint16{11, 10, 10, 21, 20, 20, 31, 30, 30, 41, 40, 40, 51, 50, 50, 61, 60, 60, 71, 70, 70, 81, 80, 80, 91, 90, 90, 101, 100, 100, 111, 110, 110, 121, 120, 120, 131, 130, 130, 141, 140, 140, 151, 150, 150, 161, 160, 160, 171, 170, 170, 181, 180, 180, 191, 190, 190, 201, 200, 200, 211, 210, 210, 221, 220, 220, 231, 230, 230, 241, 240, 240, 251, 250, 250, 261, 260, 260, 271, 270, 270, 281, 280, 280, 291, 290, 290, 301, 300, 300, 311, 310, 310, 321, 320, 320, 331, 330, 330, 341, 340, 340, 351, 350, 350, 361, 360, 360, 371, 370, 370, 381, 380, 380, 391, 390, 390, 401, 400, 400, 411, 410, 410, 421, 420, 420, 431, 430, 430, 441, 440, 440, 451, 450, 450, 461, 460, 460, 471, 470, 470, 481, 480, 480, 491, 490, 490, 501, 500, 500, 511, 510, 510, 521, 520, 520, 531, 530, 530, 541, 540, 540, 551, 550, 550, 561, 560, 560, 571, 570, 570, 581, 580, 580, 591, 590, 590, 601, 600, 600, 611, 610, 610, 621, 620, 620, 631, 630, 630, 641, 640, 640, 651, 650, 650, 661, 660, 660, 671, 670, 670, 681, 680, 680, 691, 690, 690, 701, 700, 700, 711, 710, 710, 721, 720, 720, 731, 730, 730, 741, 740, 740, 751, 750, 750, 761, 760, 760, 771, 770, 770, 781, 780, 780, 791, 790, 790, 801, 800, 800, 811, 810, 810, 821, 820, 820, 831, 830, 830, 841, 840, 840, 851, 850, 850, 861, 860, 860, 871, 870, 870, 881, 880, 880, 891, 890, 890, 901, 900, 900, 911, 910, 910, 921, 920, 920, 931, 930, 930, 941, 940, 940, 951, 950, 950, 961, 960, 960, 971, 970, 970, 981, 980, 980, 991, 990, 990, 1001, 1000, 1000, 1011, 1010, 1010, 1021, 1020, 1020, 1031, 1030, 1030, 1041, 1040, 1040, 1051, 1050, 1050, 1061, 1060, 1060, 1071, 1070, 1070, 1081, 1080, 1080, 1091, 1090, 1090, 1101, 1100, 1100, 1111, 1110, 1110, 1121, 1120, 1120, 1131, 1130, 1130, 1141, 1140, 1140, 1151, 1150, 1150, 1161, 1160, 1160, 1171, 1170, 1170, 1181, 1180, 1180, 1191, 1190, 1190, 1201, 1200, 1200, 1211, 1210, 1210, 1221, 1220, 1220, 1231, 1230, 1230, 1241, 1240, 1240, 1251, 1250, 1250, 1261, 1260, 1260, 1271, 1270, 1270, 1281, 1280, 1280, 1291, 1290, 1290, 1301, 1300, 1300, 1311, 1310, 1310, 1321, 1320, 1320, 1331, 1330, 1330, 1341, 1340, 1340, 1351, 1350, 1350, 1361, 1360, 1360, 1371, 1370, 1370, 1381, 1380, 1380, 1391, 1390, 1390, 1401, 1400, 1400, 1411, 1410, 1410, 1421, 1420, 1420, 1431, 1430, 1430, 1441, 1440, 1440, 1451, 1450, 1450, 1461, 1460, 1460, 1471, 1470, 1470, 1481, 1480, 1480, 1491, 1490, 1490, 1501, 1500, 1500, 1511, 1510, 1510, 1521, 1520, 1520, 1531, 1530, 1530, 1541, 1540, 1540, 1551, 1550, 1550, 1561, 1560, 1560, 1571, 1570, 1570, 1581, 1580, 1580, 1591, 1590, 1590, 1601, 1600, 1600, 1611, 1610, 1610, 1621, 1620, 1620, 1631, 1630, 1630, 1641, 1640, 1640, 1651, 1650, 1650, 1661, 1660, 1660, 1671, 1670, 1670, 1681, 1680, 1680, 1691, 1690, 1690, 1701, 1700, 1700, 1711, 1710, 1710, 1721, 1720, 1720, 1731, 1730, 1730, 1741, 1740, 1740, 1751, 1750, 1750, 1761, 1760, 1760, 1771, 1770, 1770, 1781, 1780, 1780, 1791, 1790, 1790, 1801, 1800, 1800, 1811, 1810, 1810, 1821, 1820, 1820, 1831, 1830, 1830, 1841, 1840, 1840, 1851, 1850, 1850, 1861, 1860, 1860, 1871, 1870, 1870, 1881, 1880, 1880, 1891, 1890, 1890, 1901, 1900, 1900, 1911, 1910, 1910, 1921, 1920, 1920, 1931, 1930, 1930, 1941, 1940, 1940, 1951, 1950, 1950, 1961, 1960, 1960, 1971, 1970, 1970, 1981, 1980, 1980, 1991, 1990, 1990, 2001, 2000, 2000}
+)
